Improved the memory debug output so it tells you the file name and line number.
Changed LB_Free to take file name and line number as arguments.
Introduced the macro LB_FREE which should normally be used
instead of LB_Free (similarly to LB_MALLOC/LB_Malloc).
Note that LB_FREE also takes care of the type cast to (void **)
so you can now be lazy and skip that...

Started writing a Multifree with variable # of arguments,
but #if'ed it out because I'm no longer sure we need it.