New Makefile system.
To build library for, say, tflop with the F90 interface, type
  gmake ZOLTAN_ARCH=tflop YES_FORTRAN=1 zoltan
To build zdrive for tflop, type
  gmake ZOLTAN_ARCH=tflop zdrive
To build zfdrive for tflop, type
  gmake ZOLTAN_ARCH=tflop zfdrive
ZOLTAN_ARCH can also be set as an environment variable, eliminating
the need to type it in each gmake command.

HTML docs are not yet updated, but the Makefiles for the individual
architectures are.

Bill:  see special variables in Makefiles/Makefile.nist* to handle
LAM vs MPICH and various Linux F90 compilers.

Makefiles for each platform are in the directory Makefiles.