New_IDs_Branch merged with main trunk of Zoltan!
The data types for IDs in Zoltan are no longer compiled into the library.
Instead, the IDs are now arrays of unsigned integers.  The number of
integer entries in an ID can be set through LB_Set_Param.  Global and local
IDs may have different numbers of entries.  The LB_ID_TYPE is currently
defined to be unsigned integers; in the F90 interface, IDs are integers.

parmetis
+ Changes to support new IDs and argument lists for Zoltan functions and query
  functions.
+ Removed LB_GID data structure from LB_Hash_Node and LB_Edge_Node; these nodes
  now point to IDs stored in other data structures.
+ In parmetis_jostle.c, changed memory error condition from
    if (comm_wgt_dim && !ewgt)
  to
    if (comm_wgt_dim && num_edges && !ewgt)
  to make oneproc cases work.
+ Changed check_graph condition for no edges in graph to MPI_Allreduce
  num_edges instead of nedges.  (nedges is the last node's number of edges;
  if it is zero, this warning would be reported.)

Still remaining to do:
+ There may be a problem with edge weights in fdriver.
+ Testing, testing, testing (any volunteers?).