Changed the return type for "compute" functions in the Group class. Now every compute object returns a boolean to indicate whether or not the calculation was successful.
The NOX::Epetra::Group currently always returns true for all of its
compute functions, but we should add error checking and return false
if a calcuation fails.

The idea behind all this error checking is that we may be able to do
some remediation when we encounter errors. For example, if a
computeRHS fails, we may be able to recover by reducing the step
length. Alternatively, it a computeNewton fails, we may be able to
revert to the gradient, and so on...