**Major** changes to configure.ac. You'll want to double check all your standard configuration options to make sure they still work. There are two major changes - the way MPI is configured and the way we hook up with Trilinos.
MPI now has a slightly different and hopefully more logicial set of
configuration options. In general, if you were using --enable-mpi,
replace that with --with-mpi-cxx. For more details, see the new file
doc/configure.html.

The changes to Trilinos were to make it slightly more logical and,
more importantly, to enable us to disable building the
examples. Apparently this was causing Alan Williams some headaches
with certain cross-compilers. Once again, the details are in
doc/configure.html.