Added more support for SGI including defining f77 and FFLAGS as well as adding complib.sgimath to the list of libraries searched for lapack functionality.
Changed XTRALDADD to CXXREPOS to be consistent with Xyce -- plus I
just liked that name better. This caused every Makefile.in to be
modified!

Also, added MOSTLYCLEANFILES = $(CXXREPOS) to all Makefile.am's that
build libraries. It may also be useful to add this to all Makefiles,
but I haven't tested this fact.