----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	build/makefile.defaults
 	packages/epetra/example/petra_nonlinear/Makefile
 	packages/epetra/src/Epetra_BLAS.cpp
 	packages/epetra/src/Epetra_BLAS_wrappers.h
 	packages/epetra/src/Epetra_LAPACK.cpp
 	packages/epetra/src/Epetra_LAPACK_wrappers.h
 Added Files:
 	packages/epetra/test/ReferenceLogs/log.12Jun2002:07:56:37
 	packages/epetra/test/ReferenceLogs/log.mpi.12Jun2002:10:15:28
 ----------------------------------------------------------------------

This update changes the way in which Fortran names are mangled in Epetra.  It is the first part of the switch to an Autotools compile and build system.  This part is being checked in to break up the Autotools update into smaller chunks.

Autotools is able to define an F77_FUNC macro that provides a platform specific name mangling scheme.  A few special cases are manually defined for now.

Until the rest of the Autotools changes are checked in, and to support the old make system for a time, a manual definition of F77_FUNC is provided.  This is used when NO_AUTOTOOLS is defined.  makefile.defults was updated to ensure that NO_AUTOTOOLS is defined.

The sytax for calling the Epetra wrapper methods (around BLAS and LAPACK functions) is unchanged.

All serial and parallel tests passed.