Added the following options for the linear solver in NOX::Epetra::Group:
"Size of Krylov Subspace" - When using restarted GMRES this sets the
size of the Krylov subspace (defaults to 300).

"Orthogonalization" - The orthogonalization routine used for the
Gram-Schmidt orthogonalization procedure in Aztec.

"Convergence Criteria" - Algorithm used to calculate the residual that is
used for determining the convergence of the linear solver.  See the
Aztec 2.1 manual for more information.

"RCM Reordering" - Enables RCM reordering in conjunction with domain
decomp incomplete factorization preconditioning.

See the doxygen documentation for more information.