----------------------------------------------------------------------
 Modified Files:
 	Epetra_CrsGraph.cpp
 ----------------------------------------------------------------------

Fix for bug 33:

Minor changes to MakeLocalIndices required for proper execution with
rectangular matrices.  I was using the DomainMap object for some
operations where the RowMap was what I should have been using.

The regression test suite caught this problem earlier this week, but
I did not see the messages (I was in too big a hurry to catch a
plane).

I ran the tests again and can *guarantee* that the test passed.