Updated to fully support coloring in parallel environments.
The additional functionality required was the addition of the connectivity
to bounding (ghost) indices since they represent distinct non-zeroes in the
local jacobian.  These bounding indices are now included in the coloring and
the resulting color map for parallel applications represents the coloring of
the columns in the original graph.