Fixed the issue with the dependencies on the header file not being picked up.
The problem was not the symbolic links (they work, makefile looks at the time stamp of the actual file, not the link), but with makedepend generating dependencies for:
src/Mesh/MeshSet.o
instead of
obj/MeshSet.o

There is no option in makedepend to get a clean job done (-p is insufficient), so I run a perl command on the make.dependencies file after it is created by makedepend.

Hopefully this will work on all platforms.