GLOBAL_PATCH functionality added.
This compiles but has NOT been tested.
I am pretty optimistic that it works though ... so feel free to try it out now.
I will test it tomorrow.

There are A LOT of performance to be gained by improving the get_next_patch function once the TSTT interface will be complete. Before doing this though, we will need the testing framework AND the tests.