Adding files for the InverseMeanRatio metric.
InverseMeanRatio and MeanRatio are now both hard-wired.  They use
element vertices as sample points, and they do ~not~ use the
compute_weighted_jacobian function.  (If we feel it is useful,
I can add the GeneralizedMeanRatio class which would not be
hard-wired.)  Both of these metrics do use the average_method
functionality, which is not consistent with the ConditionNumber
metric for the time being.  InverseMeanRatio ranges from 0 to 1,
and MeanRatio ranges from 1 to infinity.  Both have been tested
against and are consistent with the original MeanRatio.

Also, modifying the Makefile to compile InverseMeanRatio.

Fixing typos and doxygen mistakes in ConditionNumber and
AspectRatioGamma.