----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	Trilinos_config.h.in configure configure.ac
 	packages/Makefile.am packages/Makefile.in
 ----------------------------------------------------------------------
Allows Nox to be configured from the Trilinos level.  At this time,
the interaction between Nox and Trilinos is very basic.  For this reason,
the default for now will be to not build Nox.  This means that to build
Nox, the configure argument "--enable-nox" is required.

This update partially addresses bug 121.