----------------------------------------------------------------------
 Added Files:
 	AztecOO_StatusTestMaxIters.h AztecOO_StatusTestCombo.h
 	AztecOO_StatusTest.h AztecOO_StatusTestCombo.cpp
 	AztecOO_StatusTestMaxIters.cpp AztecOO_StatusType.h
 ----------------------------------------------------------------------

Partially address bug 128.

These new files are part of a suite of classes that allow user-defined stopping
criteria in AztecOO.  The base class AztecOO_StatusTest can be extended to allow
a large variety of convergence tests, and the ANDing and ORing of such tests.

The basic capabilities are working, but I am only checking in the new files while
I am doing more extensive testing.