I found myself rewriting the same chunk of code for the third time and decided to just add it to the Common object for line searches.
1. Added a set of common data outputs that all linesearches should report
and a corresponding function that can be called to set those values
in the "Output" sublist for the line search.  When performing line
search studies, it is critical to get at least this much information from
each line search for a fair comparison. See the documentation for details.

2. Added a new function called computeSlopeWithOutJac().  This is an
optimized version of computeSlope() that does not require a Jacobian.  It
uses directional derivatives.  This is needed for the soon to be checked-in
More'-Thuente2 line search.  See the documentation for details.

3. The constructor has changed slightly for NOX::LineSearch::Common so I
had to make changes to allow for this in the Polynomial, Quadratic, and
MoreThuente line searches that derive from the common object.