Modifying CG with changes described below.  Also, adding snap_vertex_to_domain  calls in a few places.
Adding a boolean return value to ObjectiveFunction's
evaluate and gradient functions.  This boolean is to
be set to false any time an ObjectiveFunction is not
defined for a given patch (e.g., an objective function
which uses the ConditionNumber metric is said to
be undefined for a patch which contains inverted
elements).  The actual objective function value is
now placed in a variable, fval, which is passed by
reference.  This is similar to the structure of the
QualityMetric's evaluate functions.  For several
tests, the code gave the same results as the previous
Mesquite version.  All unit tests pass on Sun except
test_compute_gradient_2D_LPTemplate.  This commit
should make QualityImprover::check_feasible
unnecessary, as well as the Feasible flag in
ObjectiveFunction and QualityMetric, though I have
not removed either yet.