Modifications to remove warnings when compiling with 'CC +w'. The majority of the changes are insignificant.  The main two changes:
1:  Removing 'break' statements in switches which come
after 'return' statements (and thus can not be reached).

2:  Removing variables' names  which are never used.  For example,
replacing initialize(PatchData &pd, MsqError &err) with
initialize(PatchData &pd, MsqError &/*err*/) if the MsqError
object is not used in the initialize function.

Modification 2 is slightly inconvenient, please let me know if
it causes any problems.  However, I think it would be helpful to
be able to compile with all the warning flags enabled.