Added a NOX Random class to encapsulate the calculation of random numbers. The algorithm used to compute random numbers is the same as that used by Epetra and relies on the POSIX rand() system call to compute a seed number.
Added methods to the NOX::Abstract::Vector class to fill a vector with random
values and set the random number seed.  These methods have default
implementations that print an error message and throw an error.

Added an autoconf test to check for cstdlib and stdlib.h and include them in
NOX_Common.H