Added a LOCA::Predictor::Random class to compute a predictor whose solution component is filled with random values.  This is used primarily for computing curves of turning points where the Jacobian is singular.
Modified the secant predictor to be able to use any other predictor for the
first step (when there isn't a previous point to calculate a secant vector
from).  Typically, the first step predictor is either the constant predictor
or the random predictor.  The latter allows arc-length continuation of
turning points using the secant predictor.

Fixed the scaling of the secant predictor so arc-length continuation using
the secant predictor works for poorly scaled parameters.

Added an applyJacobianInverseMulti() method to the LOCA::Abstract::Group to
compute solutions for multiple right-hand sides.  It has a default
implementation which just computes the solutions sequentially.  Also added
multi-versions of the Nic, NicDay, and ItRef singular solve methods.

Implemented applyJacobianInverseMulti for turning point groups and implemented
the arc-length applyJacobianInverse in terms of the underlying
applyJacobianInverseMulti.

I'm sure there's more here that I've forgotten...