----------------------------------------------------------------------
 Modified Files:
 	Ifpack_OverlapGraph.cpp
 ----------------------------------------------------------------------

Closes bug 214.

Background:

When calling TransformToLocal for an overlapped graph, we pass the domain map of
the original non-overlap graph in as the domain map for the overlap graph.  This
gives us the automatically correct importer for getting vector elements for the
overlapped solve.  Unfortunately, use of this domain map makes column IDs that
should be considered local to be considered remote and the column map ordering
of the overlap graph is inconsistent with the row map ordering.

Because of this inconsistency, the factorization (which is based on local IDs
only) does not see a symmetric permutation of the rows and columns and elements
which should be diagonal are not.

To fix this, when the last iteration of overlap is computed the overlap graph
is now constructed by specifying both the row map and the column map
explicitly and making them the same. This then forces a consistent ordering.