----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	doc/DevGuide/TrilinosDevGuide.tex
 	packages/new_package/aclocal.m4 packages/new_package/configure
 	packages/new_package/configure.ac
 	packages/new_package/src/New_Package_config.h.in
 	testharness/HowToAddToTestHarness
 ----------------------------------------------------------------------
Partially address bugs 179, and 13.

Latest (small) changes to developer's guide.

Added a little bit of detail to the HowToAddToTestHarness file.  It is
still not complete.

Cleaned up and commented the configure.ac file for new_package quite a bit.
This file is probably the single most important file to do a good job on in
new_package.

Tests pass.