----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	packages/aztecoo/Makefile.in packages/aztecoo/aclocal.m4
 	packages/aztecoo/configure packages/aztecoo/configure.ac
 	packages/aztecoo/example/Makefile.in
 	packages/aztecoo/example/AztecOO/Makefile.in
 	packages/aztecoo/example/AztecOO_UserOpUserMat/Makefile.in
 	packages/aztecoo/example/simple_epetra_aztecoo/Makefile.in
 	packages/aztecoo/src/Makefile.am
 	packages/aztecoo/src/Makefile.in
 	packages/aztecoo/test/Makefile.in
 	packages/aztecoo/test/AztecOO_UserOpUserMat/Makefile.in
 	packages/aztecoo/test/scripts/Makefile.in
 	packages/aztecoo/test/scripts/daily/Makefile.in
 	packages/aztecoo/test/scripts/daily/serial/Makefile.in
 ----------------------------------------------------------------------
Partially addresses bug 252.

AztecOO began to have problems with templates on CPLANT after a recent
checkin.  At the suggestion of Rob H, Trilinos will support one general
configure line option, TRILINOS_TEMPLATE_OBJS=, which will be recognized
by all packages that have template issues.  The old method that was used
to address the problem of templates on cplant was to use package specific
options, ie EPETRA_TEMPLATE_OBJS.  This was chosen before we knew that
3-4 packages would require such an option, rather than just 1.  This commit
does not close out this bug; it only addresses it for AztecOO.

Tests pass.