Added a new Status Test called NOX::StatusTest:Stagnation.  This has been requested by John Shadid, Homer Walker, and the Xyce team for various projects over the last few months.
The user can specify a tolerance, t,  and a number of iterations, n.  If the
convergence rate is above t for n consecutive iterations, the status test
will return a Failed status.

This is used to prevent programs from wasting time on
computations that are going nowhere but are still running since they have
not yet hit the maximum number of iterations (especially important for
line searches).

This has prevented failed runs in MPSalsa from taking 24 hours to complete.
It realizes things are bad and stops in only a few hours.