getting rid of writeVtkMesh etc ....
Those used the old TSTT interface, they are not needed with Darryl's new implementation. Those functions had to deal with non-persistent TSTT handles and were painfully slow as a result. We'll be better off rewriting them for scratch for the new TSTT interface. Or we can always dig them up from CVS. In any case, they should not be in Mesquite right now.