----------------------------------------------------------------------
Committing in .

Modified Files:
	epetra/test/BlockMap/cxx_main.cpp
----------------------------------------------------------------------
Addresses bug 384.

Test formerly attempted to cast a pointer back to an int.  On SGI64 this
failed because the pointer is 64 bit and the int 32 bit.  At the suggestion
of Mike H., we are now using size_t instead of int.

MPI and Serial linux tests pass.