----------------------------------------------------------------------  Modified Files:  	packages/epetra/src/Epetra_IntSerialDenseMatrix.h  	packages/epetra/src/Epetra_IntSerialDenseVector.cpp  	packages/epetra/src/Epetra_IntSerialDenseVector.h  	packages/epetra/src/Epetra_SerialDenseMatrix.h  	packages/epetra/src/Epetra_SerialDenseVector.cpp  	packages/epetra/src/Epetra_SerialDenseVector.h  ----------------------------------------------------------------------
IntSerialDenseVector and SerialDenseVector did not have an assignment operator defined, and because operator= is not inherited, they were using the compiler-generated versions, which produce the Wrong behavior. I have added an overloaded operator= to both IntSerialDenseVector and SerialDenseVector. Because IntSerialDenseVector and SerialDenseVector are just trivial cases of IntSerialDenseMatrix and SerialDenseMatrix, all the assignment operators do is call op= on their base class portion.

This should resolve bug 388.

I also corrected a typo in the documentation of IntSerialDenseMatrix and SerialDenseMatrix.

All Epetra tests pass (linux serial and linux mpi).