----------------------------------------------------------------------  Modified Files:  	packages/epetra/src/Epetra_IntSerialDenseMatrix.cpp  	packages/epetra/src/Epetra_IntSerialDenseMatrix.h  	packages/epetra/src/Epetra_IntSerialDenseVector.cpp  	packages/epetra/src/Epetra_IntSerialDenseVector.h  	packages/epetra/src/Epetra_SerialDenseMatrix.cpp  	packages/epetra/src/Epetra_SerialDenseMatrix.h  	packages/epetra/src/Epetra_SerialDenseVector.cpp  	packages/epetra/src/Epetra_SerialDenseVector.h  	packages/epetra/test/SerialDense/cxx_main.cpp  ----------------------------------------------------------------------
A new constructor has been added to (Int)SerialDense{Matrix,Vector}. It takes the dimensions of the matrix/vector as parameters, and sizes itself accordingly. Calling this constructor is equivalent to using the default constructor, and then calling either Shape (in the case of a Matrix) or Size (in the case of a Vector).

SerialDenseTest has been modified to call the new constructor in SerialDenseMatrix and SerialDenseVector, and check that they were constructed correctly. (This capability will be a part of IntSerialDenseTest when it is added to the repository).

All Epetra tests pass (linux mpi, linux serial).