----------------------------------------------------------------------
Committing in .

Modified Files:
	test-harness
----------------------------------------------------------------------
Partially addresses bug 330.

Copied a fix that Ken S. applied to the checkin-test-harness.  Problems
occured when a script did not return us to the directory that we entered the
script in and there were multiple scripts in one directory to run.

This will hopefully allow dscpack tests to pass for the nightly Solaris, Linux,
and SGI tests.