----------------------------------------------------------------------  Modified Files:  	config/AUTOTOOLS_FAQ  Added Files:  	config/macosx_mpi_paul config/macosx_serial_paul  ----------------------------------------------------------------------
I have successfully gotten Trilinos to build in both serial and mpi
mode on my Quicksilver G4, running MacOS X 10.2.6. All of the Trilinos
packages build, with the exception of amesos. (And that seems to be
due to other components that amesos requires that I haven't configured
properly, and not related to OS X or to the config script. But as I'm
not an amesos developer, nor an amesos user, it's not a big concern of
mine.) As an added bonus, these scripts should be able to take
advantage of the Altivec-optimized BLAS and LAPACK libraries in
Jaguar's vecLib framework.

I also corrected a typo in the AUTOTOOLS_FAQ file.

All Epetra tests pass (OS X serial, OS X mpi).