More changes to the mesh writing/reading functions.
Making write_exodus write a nodeset corresponding to the 'fixed'
vertices in the vtk file.  The nodeset is arbitrarily numbered 111.
Also, making read_exodus look for this nodeset and mark the
corresponding vertices.

Adding ex_close(...) calls to both read_exodus and write_exodus.

Making write_vtk(...) print more significant digits so that less information
is lost when writing to a file.

Passes Unit tests with and without Exodus enabled.