Changed how status tests are passed to the stepper.  Previously, there were two constructors that took either 1 or 2 status test arguments.  If 1 was supplied, it was assumed that status test would work for both the original group and the continuation super-group (e.g., NormF), and if 2 were supplied, the first was for the original group and the second was for the parameter component of the continuation super-group.  They were then and'd together after the first nonlinear solve.
Now there is only one constructor, taking one status test, which is assumed
to be appropriate for the entire continuation group.  The first nonlinear
solve is now done with a continuation group as opposed to the original group,
so there is no difference between the first step and subsequent steps.  The
status test can be either something that isn't group specific, i.e., NormF
and MaxIters, or can be some composite using the the LOCA status tests.