Added a new status test to check for Inf and NaN results.  This was needed for Xyce and MPSalsa and I think Russ Hooper mentioned wanting it for SIERRA some time ago.  Since so many codes asked forthis, it makes sense to add it to nox for general use.
For transient and continuation runs if the time/parameter step is too
big, the nonlinear solver could diverge and create unphysical solution
values.  The resulting computeF() call would return a vector filled
with NaNs of Infs.  Our status tests were actually claiming
convergence in some cases when the norm of F = NaN.  This check looks
at the norm of F and will return a failure if it detects NaN or Inf.

I have added a demonstration of it's use in 1Dfem_nonlinear.