----------------------------------------------------------------------
Committing in .

Modified Files:
	TrilinosDUGuideConfigBuild.tex
----------------------------------------------------------------------
Small contribution to bug 561.

Added a description of the --enable(disable)-tests and
--enable(disable)-examples options.  I will have to highlight the use
of these somewhere else yet.

Added a key point noting that it is worth figuring out exactly which
packages should be built and enabling and disabling as necessar because it
takes a long time to configure and build.