----------------------------------------------------------------------  Modified Files:  	aclocal.m4 configure configure.ac example/Makefile.am  	example/Makefile.in src/Makefile.am src/Makefile.in  	src/Tpetra_ElementSpace.cpp src/Tpetra_ElementSpace.hpp  	test/Makefile.am  Added Files:  	example/verySimple/Makefile.am example/verySimple/Makefile.in  	example/verySimple/cxx_main.cpp src/Tpetra_Object.cpp  	src/Tpetra_Object.hpp src/Tpetra_config.h.in test/Makefile.in  ----------------------------------------------------------------------
Added a "verySimple" example directory. It creates a Tpetra::Object, and calls a member function on it.

Tpetra successfully builds with autotools. However, at present, only Tpetra_Object is integrated into the build process, along with the verySimple example.

(Tpetra is not yet integrated into the Trilinos build system. I will be checking that in as soon as I build everything else & run the testharness to confirm that adding Tpetra didn't break anything.)