All TSTT unit test now pass with AOMD.
Unfortunately, I had to catter to AOMD implementation of Tags, which is probably different from the TSTT standard. When AOMD sets/gets a TSTT tag data, it in fact sets/gets a pointer to that data. The space for the data is supposed to be managed by the application (mesquite) at all times.

We have to check if all TSTT implementations understand tags in that manner.