Added
#ifndef HAVE_CONFIG_H
#define HAVE_CONFIG_H
#endif

to the list preprocessor code passed in the AC_LANG_SOURCE calls in the
TRILINOS_HEADER macro in an attempt to fix bug #689 where configure is failing
on the SGI when trying to preprocess Teuchos headers.  The problem appears to
be stemming from the fact that HAVE_CONFIG_H is not defined when processing
headers from other packages, and therefore Teuchos_ConfigDefs.hpp is
relying on fallback code that is apparently incorrect for this particular SGI.

This points to the more general problem that when configure trys to
process headers from other packages (and also when we try to compile our
code that uses those headers) we really need all of the C, CXX, CPP, and F77
flags from those packages in our flags to ensure proper compilation.