----------------------------------------------------------------------
 Modified Files:
 	icrout_cholesky_mex.c
 ----------------------------------------------------------------------

Fixes bug 717.

Ulrich Hetmanyuk identified the following bug:

In the file "icrout_cholesky_mex.c", in the routine "crout_ict",
the line 382 is

norm_l /= (double) len_l;

When len_l is 0, this line generates a NaN for norm_l.

This NaN can create a floating exception (for instance on Stratus).

With GNU compilers, the code continues to run because
the value of norm_l is not referenced when len_l is 0.

A more robust division (without creating NaN) should be used.
For instance, we may use

norm_l = (len_l == 0) ? 0.0 : norm_l/((double) len_l);

I made the recommended change and tested it.