----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	Makefile.am Makefile.in aclocal.m4 configure configure.ac
 	doc/Makefile.in doc/Doxygen/Makefile.in src/Makefile.in
 	src/TSF_config.h.in src/Concrete/TSFEpetraVectorType.cpp
 	test/Makefile.in test/Double/Makefile.am
 	test/Double/Makefile.in test/PoissonTest/Makefile.am
 	test/PoissonTest/Makefile.in test/RectMatrixTest/Makefile.am
 	test/RectMatrixTest/Makefile.in test/VectorTests/Makefile.am
 	test/VectorTests/Makefile.in
 ----------------------------------------------------------------------

I cleaned up the configure.ac file to customize it a little more for
TSFExtended.

Added the option to disable tests.  This required changes to configure.ac
and to Makefile.am.

Reordered the link lines for tests as described in bug 613.

Put ifdef's around a #include Epetra_MPIComm.h statement to allow for
serial builds.

Eventually I got a serial build to compile without any errors.  I will
be committing another change to improve the dependency detection for
TSFExtended.