----------------------------------------------------------------------  Modified Files:  	packages/kokkos/src/Kokkos_Permutation.hpp  	packages/tpetra/src/Tpetra_CisMatrix.hpp  	packages/tpetra/src/Tpetra_Vector.hpp  	packages/tpetra/src/Tpetra_VectorData.hpp  	packages/tpetra/test/CisMatrix/cxx_main.cpp  ----------------------------------------------------------------------
Added a #include for Kokkos_Vector and Kokkos_MultiVector in Kokkos_Permutation.

Changed Vector's constructor so that the STL vector is initialized to zeros in its constructor, instead of in a loop in the Vector ctr.

More work on CisMatrix. Fixed a nasty allocation bug. Everything *should* be working now.