Provide a fix for Bug #794.
The loop over the number of colors allowed processors with fewer colors to
become out-of-sync with those having more.  Now, the max number of colors on
all processors is found and used to construct the loop.  Also, a flag
is introduced to indicate whether or not a color is valid for a processor
and therefore whether or not to contribute entries into the numerically
computed Jacobian.