1. Added timing functions and accessors to the Aztec Linear system for the preconditioner and linear solve.
I have not added similar functionality to the EpetraNew::Group because
it would be a mess.  Multiple groups are floating around (we would
have to sum all group timings together) and if they get wrapped by
loca it really gets complex.  The only way to do this would be to use
parameters lists which becomes too expensive in certain cases (like
xyce).

2. Rearranged some ctor arguments to remove some compiler warnings.