Introduce a new option: --with-amesos-mpi-c2f     (default is yes) to enable functions to convert a C communicator into a FORTRAN communicator.
This is a non-standard task, and probably the best way to deal with it
is to allow users to enable/disable those lines of code. I will add a
function to use user's provided communicators if it is not possible to
convert from C to FORTRAN communicators.  It is used in MUMPS.