Several changes: 1) Renamed LOCA::ExtendedVector to LOCA::Extended::Vector to be consistent with previous naming conventions and the new LOCA::Extended::MultiVector.
2) First implementation of LOCA::Extended::MultiVector, which is the multi-
vector version of LOCA::Extended::Vector.  The code seems to compile on linux
but has not been tested yet.  I believe this will also fix bug #804, but I
haven't tested that yet.

3) Change to Roger's fix regarding salsa and the "Singular Solve" sublist
using a const parameter list.  I took out the const_cast<> and removed
the const parameter list version of the constructor and replaced it by
a default constructor.