----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	configure configure.ac
 ----------------------------------------------------------------------
Bug 703.

We will be moving to a more flexible way of specifying a value
for TRILINOS_TEMPLATE_OBJS.  The old way set a value for a particular
case and allowed users to specify one in other cases.  The new way will
allow a user to set the value at any time, but will still try to detect
situations where a value is necessary.