----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	Makefile.in configure configure.ac example/Makefile.in
 	example/Nonlin/Makefile.in example/Nonlin/NP2DSim/Makefile.in
 	example/Nonlin/NP4DOpt/Makefile.in example/Solvers/Makefile.in
 	src/Makefile.am src/Makefile.in test/Makefile.in
 	test/epetra-adapters/Makefile.in test/scripts/Makefile.in
 	test/scripts/daily/Makefile.in
 	test/scripts/daily/mpi/Makefile.in
 	test/scripts/daily/serial/Makefile.in
 	test/scripts/weekly/Makefile.in
 	test/scripts/weekly/mpi/Makefile.in
 	test/scripts/weekly/serial/Makefile.in
---------------------------------------------------------------------
Closes Bug 703.

We are moving to a more flexible way of specifying a value
for TRILINOS_TEMPLATE_OBJS. The old way set a value for a particular
case and allowed users to specify one in other cases. The new way will
allow a user to set the value at any time, but will still try to detect
situations where a value is necessary.

TSFCore uses TRILINOS_TEMPLATE_OBJS only conditionally.  I set up
an AM_CONDITIONAL value to recognize when we need TRILINOS_TEMPLATE_OBJS.
This conditional is now checked in the src/Makefile.am before including
the path.