----------------------------------------------------------------------
 Modified Files:
 	test/inout/cxx_main.cpp
 	test/scripts/daily/mpi/autotoolsTestAllMpi
 	test/scripts/daily/serial/autotoolsTestAll
 ----------------------------------------------------------------------

Modified the list of directories specified to the foreach command so
that only the directories that have working test codes are listed.
This revealed a problem with the verbosity of the inout test, which I
also fixed.  There is another problem with Permutation, which I was not
able to track down in the time remaining today.  I will file a bug report on
it.