----------------------------------------------------------------------
 Modified Files:
 	Epetra_BlockMap.cpp
 ----------------------------------------------------------------------

Fixes bug 868.

This is the same fix as what was entered in the Release 4 branch.
Checking the history of this change with Bonsai, it turns out that
this is a fairly new bug introduced with some optimizations that
Alan Williams provided.  I don't think my fix will have any major impact
on the overall timing results of his work.

This bug was not part of Trilinos 3.1.