----------------------------------------------------------------------  Modified Files:  	src/Tpetra_CisMatrix.hpp src/Tpetra_Util.hpp  	test/BlockElementSpace/cxx_main.cpp  	test/CisMatrix/cxx_main.cpp test/Distributor/cxx_main.cpp  	test/ElementSpace/cxx_main.cpp test/Platform/cxx_main.cpp  	test/Vector/cxx_main.cpp test/VectorSpace/cxx_main.cpp  ----------------------------------------------------------------------
The tests for BlockElementSpace, CisMatrix, Distributor, ElementSpace, Platform, Vector, and VectorSpace have been adapted to run in either serial or MPI mode. (Although CisMatrix and Vector seem to be crashing when run in MPI at the moment.)

Added an efficientAddOrUpdate function to Tpetra_Util.

Added cases for Add and Replace CombineModes to CisMatrix.submitEntries.