Fix to work with standard CppUnit.
Removed includes of non-existent cppunit/SignalException.h .
MesquiteTestRunner depends on non-existent "key" for TestSuites.
  - use text-ui TestRunner included with CppUnit instead
  - don't compile MesquiteTestRunner

I couldn't find any reference to the above stuff in either the current
or older versions of CppUnit.  The sf.net site for CppUnit doesn't have
a beta/cvs version newer than the current release (1.10), so I'm
assuming the above stuff is either someone's customized version of
CppUnit or a beta/cvs feature that was never included in an offical
release of CppUnit.