Added a class for general additive Schwarz preconditioners. This class handles the overlapping and the non-overlapping case. The class is based on Epetra_RowMatrix.  The overlapping matrix must be furnished by the user (class Ifpack_CrsAdditiveSchwarz, for Crs matrices only, can automatically build the overlapping matrix).
This class is templated with the local solver, which must be a
Ifpack_Preconditioner derived object. It is understood that the local
problem will be solved by using the ApplyInverse() of this object. Using
templates allow quite a generality in the code, as essentially every
preconditioner that can be wrapped ad Ipfack_Preconditioner can be used.