----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	src/Epetra_CrsMatrix.cpp test/CrsMatrix/cxx_main.cpp
 ----------------------------------------------------------------------
bug 966.

Fixed invRowSums() method so that it works if partial sums are contributed
by multiple processors and when rows are replicated across multiple
processors.  I still have to document exactly how this works, but I wanted
to get the code checked in.

I also wrote a couple of tests for the method.  The Epetra_CrsMatrix tests
pass (serial and mpi).