----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	configure configure.ac
 	config/tac_arg_enable_default_feature.m4
 	config/tac_arg_enable_feature.m4
 ----------------------------------------------------------------------
bug 991.

Configure will now pass --enable-package_name down to the individual
package configure scripts even if package package_name was not explicitly
enabled (ie package_name is a default package or another package that was
enabled depends on package_name).

This will be useful because individual packages will no longer have to
"know" which packages are default packages and will not have to worry
about whether or not --disable-default-packages was used.  We could
even change the set of default packages without any necessary changes
at the individual package level.