OverlappingRowMatrix is a class, derived from Epetra_RowMatrix, that allows the construction of an overlapping matrix, by importing only the overlapping rows. In its actual implementation, the class contains a pointer to the non-overlapping matrix, plus a pointer to an Epetra_CrsMatrix, contining *only* the overlapping rows. This saves a considerable amount of memory.
Ifpack_vRiluk is a value-based RILUK, and it is an alternative the
Ifpack_gRiluk. In this new class, first the algorithm computes the row `i' of
the incomplete factors L and U as in a complete factorization. Then, only
the elements that are in the sparsity pattern of A, and the largest
LevelOfFill() elements are kept. If LevelOfFill() == 0, vRiluk requires
less memory than gRiluk (no graph is required).