----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	daily/mpi/epetraTestAllMpi daily/serial/epetraTestAll
 ----------------------------------------------------------------------
Bug 1037.

Test failures have been reported in email attachments, but the subject
lines of the emails have been reporting success in all cases.  This was
due to the fact that the test scripts have not been properly returning
the error code.  I am not sure how failures were properly reported in the
past given what was wrong.  I tested this fix with tests that fail and
tests that pass to assure that both cases work.  I still need to apply
this fix to other packages.