----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	TrilinosUserGuide.tex
 ----------------------------------------------------------------------
Modified the instructions for how to test a build.  Noted that we are
working on a way that users will be able to run tests to verify that a
build was successful.  Also gave a pointer to test harness instructions
on the web and provided step-by-step instructions for running individual
tests.