Modified MatrixMatrix::Multiply to address a bug discovered by Roman Geus.
Modified EpetraExt_BlockMapIn to add a function that creates 4 maps (row,
col, range and domain) from a matrix-market matrix file. This allows a
test program to correctly read a distributed, rectangular matrix from a
matrix-market file that wasn't produced by Epetra or EpetraExt.

Modified EpetraExt_CrsMatrixIn to add a function that reads a matrix from
a matrix-market matrix file, allowing the user to specify the 4 maps to
be used for matrix construction and for calling matrix.FillComplete().
Again, this allows a test program to read a rectangular matrix into a
parallel test run.

Modified epetraext/test/MatrixMatrix to use the inout functions for
reading test matrices from matrix-market files. The test program will
be completely input-driven, reading specified matrix-files to use as
matrix-matrix product operands, and also reading a specified matrix-file
to check the result against. This allows a single test program to exercise
many paths through the matrix-matrix product code, providing a much
better test.

This test program is not yet active, another bootstrap and cvs commit is
necessary to get test input files in the correct place.

Added several matrix-market files to the test/MatrixMatrix directory, to
be used for test inputs and for checking test results.