Adding a class that can take a MeshSet and perform an affine transformation on it.  This is another PatchDataUser (like QualityAssessor or QualityImprover).
I'm not adding tests for it, yet.  I want to make sure this is the type of
thing that we are wanting.  I will add a driver code shortly.  If we decide
to expand this type of capability, then I'll move this into another directory,
but I don't really expect that it will be expanded too far.

Also, adding the ability to create an ELEMENTS_ON_VERTEX_PATCH with an element
depth of 0 (create a PatchData object that contains a single vertex).  I also
am not adding Unit tests for this, yet.

Passes Unit and Regression on Sun.