Updated some of the doxygen documentation in Epetra_CrsMatrix.h and in Epetra_CrsGraph.h. Neither of these are finished yet, more updates to come. Partially addresses bug 412.
Added a check in Epetra_CrsMatrix to detect whether FillComplete was
called on the matrix' graph attribute after the matrix was constructed
and before FillComplete was called on the matrix. If this happens (for
instance if multiple matrices share a graph that hasn't been Filled())
then the matrix data could be out of sync with the graph and
matrix.FillComplete() may not be able to complete without errors.
If this situation is detected a warning message is printed but no drastic
action (such as a throw) is taken. It is possible that the user knows
what they're doing and data is not out of sync, but this is a dangerous
situation.

Further investigation (and a better solution) is required, but this
will address an issue raised by bug 708.

All epetra tests pass.