Committing the swig generated wrapper files for the NOX/LOCA-python interface, and the configure/make changes required to support this.  This *should* make it possible to build the python interface without swig.  To turn on the ability to create these wrapper files using swig, configure with --with-swig, which can take an optional argument giving the swig binary you wish to use (in case you want to use a different swig binary than the one first in your path).
Some basic dependency tracking is available for swig as well, and it works
in a very similar manner as automake's dependency tracking.  If you remove
the swig wrappers with 'make swigclean' and recreate them using 'make all',
you'll see how it works.

Because of some added automake conditionals and AC_SUBST's, every Makefile.in
was touched by this commit.