----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	packages/nox/test/scripts/daily/mpi/noxTestAllMpi.in
 ----------------------------------------------------------------------
Changed "if(! -z ../$file4 ) then" to "if(! -z ../../$file4 ) then"
When the script was adapted from the Epetra version, this minor
change must have been overlooked.  Fixing the problem on herouxsmp
has allowed the NOX tests to pass for the first time in a very long
time.  I sent Roger a note about this change for approval.