Fix for bug 1079.
Replaced all instances of max and min with NOX_MAX and NOX_MIN macros.  This follows what other Trilinos packages have done.  The problem is that some compilers don't define max/min while others do.  nox was undefining compiler versions and using our own define, but apparently this does not work for all platforms.

The offending code is shown below with the comment as to why it was added.  All instances of this block have now been removed.

/* Some compilers (in particular the SGI and ASCI Red - TFLOP)
 * fail to find the max and min function.  Therfore we redefine them
 * here.
 */
#ifdef max
#undef max
#endif

#define max(a,b) ((a)>(b)) ? (a) : (b);

#ifdef min
#undef min
#endif

#define min(a,b) ((a)<(b)) ? (a) : (b);