Fixed a couple of issues with memory management in Aztec that were causing errors in some Alegra tests.
These problems originated with the changes made in December to allow the
AZ_iterate function to be used recursively. The side-effect was that if
multiple matrices are created and solved separately (with no recursion),
setting AZ_reuse for one but not the other, then when AZ_free_memory was
called it could clean up memory for the wrong matrix if data_org[AZ_name]
values were a certain unlucky set of values. This problem has been fixed.
The logic that handles recursive usage of AZ_iterate is now more robust
and can't effect non-recursive use-cases.

The test program in test/AztecOO is also now more extensive.

purify has been used to verify that Aztec runs clean and without leaks,
in the test program as well as in the Alegra cases.