A few things
1) I made a minor change in how the perl script .generate-makeoptions.pl is called.  Technically this does not really change anything but it is a little cleaner.

2) Added a new Describable interface and added an entry into the doxygen index.doc file.
This interface will be the base class for all new Thyra interfaces and Teuchos seemed like the right place to put this.  I still may need to include an example/testing program to demonstrate how this interface it used.  However, this interface is pretty simple and may not need much more documentation other that what is already provided in the class.