Fixing yet another portability issue with my previous commit.
The cbrt() function is defined in the C99 standard but not in any
C++ standard.  This means that on some platforms it is available in
the <math.h> header, but not <cmath>.  The configure script decides
to use the <cmath> header if it is available, because that's what
the C++ standard says we should be using.

To strictly conform to the C++ standard, we cannot use cbrt().  I
don't think that's an option.  I think the best solution is probably
to ignore the C++ standard with regard to C headers, and just use
the .h forms for all C headers (not C++ headers.)

However, this hack seems to resolve the issue for both Sun and
Linux without changing all the code to use C headers.  It just
includes both <cmath> and <math.h>.  If this solution causes
problems on some other platform in the future, I'll just change
the code to use .h C headers everywhere.