Small change to make I_DFT_NoBarrier work on inverted meshes even if useBarrierDelta is false.
Another option here is to go into I_DFTFamilyFunctions and do some additional
if checks for gamma==0.  We can save a few flops by doing that, but it
will make I_DFTFamilyFunctions a little more complicated, and I'm not sure
whether it would a over-all speedup (... trading flops for if-checks).

Passes Unit on Sun.