I'm starting to get the abstract interfaces defined.
  The NonlinearModel class is an abstract class that uses InArgs and OutArgs to
define how to evaluate the model.  ExampleApplicationRythmosInterface is now derived from NonlinearModel and implements the necessary functions in terms of epetra vectors.

  The Stepper class takes as input a NonlinearModel object which it uses to
evaluate the model.  The Stepper_ForwardEuler class is derived from Stepper and
implements forward Euler in terms of Thyra vectors.

  I also implemented the beginnings of templating on the Scalar type.  This
does not propagate to the ExampleApplication code (as is appropriate), but I
think I did everything right for the Rythmos code.

  After talking with Heidi, I've decided I should use MultiVectorBase as my
workhorse and assume everything coming in and going out is of that type.  This
necessitates some additional work on my part, which I'm not quite to yet.