o Change algorithm from ConjugateGradient to FeasibleNewton.
o Test target calculator and DFT metrics.

o Reduce amount of deformation so metrics don't fail.

This test passes *if* numerical Hessians are used for the mean
ratio metrics.  However, it is still failing with for the mean
ratio metrics with analytic hessians.  The I_DFT metrics appear
to work correctly.