Do not include test_TSTT in "all_tests" build target unless Mesquite was configured with a TSTT Mesh server (e.g. MOAB) against which to test.
If Mesquite is configured w/out CPPUnit support, do not include the
unit tests in the "all_tests" build target.

If Mesquite is configured w/out CPPUnit support, make the "test" target
print a message to that effect rather than pages of link errors and
potentially running an old version of the unit tester.