Address bug 857. The Epetra_Directory interface (and the Epetra_BasicDirectory implementation thereof) now has a query function 'bool GIDsAllUniquelyOwned()' which returns true if all GIDs appear on only one processor, or false if any GID appears on more than one processor.
A couple of new data attributes were added to Epetra_BasicDirectory:
A list-of-lists to hold a list of processors for each directory entry,
and a corresponding array of list-lengths.
These arrays are only allocated if one or more directory entries has
multiple owning processors. Even then, no list is allocated for directory
entries that only have 1 owning processor.

For directory entries with just 1 owning processor, the ProcList_
array is still used as it was before. For entries with multiple owning
processors, the ProcList_ array holds the lowest-numbered of the
owning processors. This can be enriched later to account for schemes
that choose the owning processor to be something other than the
lowest-numbered one.

We will also want to add more query methods to the Epetra_Directory
interface to do things like return a list of GIDs that are shared
by multiple processors, and/or return a list of processors for a
given GID.

epetra/test/definition has been edited to add the Directory test.

All epetra tests pass, and the Directory test has been checked with
purify.