Implemented the first version of adaptive balance adjustment during recursive bisection. My limited tests showed better results in terms of cut quality.
I've also cleaned up some of the code/interfaces. We should get rid of
hg->ratio; or at least we should not use it in PHG. Multiple reasons
for this, first hg is not an appropriate place for that information,
second ratio is only useful for bisection; instead we're using
part_sizes.