----------------------------------------------------------------------
 Modified Files:
 	src/Epetra_CrsMatrix.h test/CrsMatrix/cxx_main.cpp
 ----------------------------------------------------------------------

Fixes bug 1490.

Added a method to extract raw pointers from an Epetra matrix that has
been fill completed and storage optimized.  This method is part of
an "expert-only" section and is necessary for optimal performance of
Gauss-Seidel type preconditioners.

Added a test for the method to CrsMatrix test driver.

All test run and passed.