----------------------------------------------------------------------  Modified Files:  	configure configure.ac config/replace-install-prefix.pl  	test/Makefile.am test/Makefile.in test/definition  Added Files:  	test/GMRES_Tests/Makefile.am test/GMRES_Tests/Makefile.in  	test/GMRES_Tests/cxx_main.cpp  ----------------------------------------------------------------------
Fixes bugs 1722 and 1723 in development branch.

Modified thresholds to be more reasonable, especially for situations
where there is a large magnitude difference in values in the residual
vs. the matrix coefficients.

Also added the ability to choose between one and two steps of classical
or modified Gram-Schmidt orthogonalization.

Added a new test for az_gmres that builds a 5x5 diagonal matrix that
has separated diagonal values with only one positive value.  This
forces exactly 5 itertions and allows us to observe the behavior of the
four orthogonalization methods.