Modified the tutorial with a new style, which is much nicer than before. The document looks "younger"; the content is exactly the same at this point.
Also, I put the original three authors as "Editors", then each chapter reports
the corresponding authors in the chapter. This model is more flexible and
localizes the contributions of each chapter.

Also, the makefile now uses pdflatex, which produced better PDF files. This
required a conversion of DOEbwlogo and snllineblk to PDF format.