Several changes:
1)  Added preProcessContinuationStep() and postProcessContinuationStep()
methods to the group and constraint hierarchy which are methods that are
called before and after each continuation step.  They take the step status
as an argument, so the group can implement behavior depending on whether
the step succeeded or failed.  This replaces the notifyCompletedStep()
method, which was only called after a successful step.  Changed all of the
groups and constraints to use these methods when appropriate.

2)  Added a modified constraint method for the minimally augmented formulation
that solves for the updates to the null vectors for each nonlinear iteration
instead of solving directly for the null vectors.  This requires some
new derivatives which were added to DerivUtils.

3)  Cleaned up the minimally augmented turning point formulation considerably.