A first commit for initial Zoltan partitioning capability in Isorropia. At this stage, Zoltan partitioning/rebalancing is only supported for Epetra_CrsGraph objects.
The code in Isorropia_Zoltan_Rebalance.cpp is essentially pulled straight
from Rob Hoekstra's code in
 epetraext/src/zoltan/EpetraExt_Zoltan_CrsGraph.cpp.
It utilizes code that is still in epetraext/src/zoltan, such as
Zoltan_LoadBalance, etc.

To enable Isorropia_Zoltan, it is currently necessary to specify
--enable-epetraext-zoltan in the configure command. We will decide
whether to pull all epetraext/src/zoltan code into Isorropia so that
the extra --enable clause is not necessary.

This capability is passing an initial test, in
 test/zoltan/rebalance_zoltan.cpp which I've run on 1,2,3 processors.

The capability is only available through Isorropia's create_balanced_copy
function, which provides the same functionality as the Transform object
in EpetraExt was providing.

More refinement is necessary, including more parameter parsing, etc.

We also will now proceed with defining Isorropia interfaces for
Partitioner, Migrator, and 'cost describer'.