Added a setParameters method to the Partitioner interface so that parameters can be set after construction.
Also made it so that the paramter list passed to both the constructor
and this new method is of type 'const Teuchos::ParameterList&'. When it
was wrapped in a Teuchos::RefCountPtr (previously), it caused bad semantics
because if a user gave a parameter-list to the partitioner and then
changed the contents of the parameter-list, those changes would be
automatically reflected in the partitioner's parameter-list.
Now, the partitioner takes a copy of the parameter-list, and is therefore
isolated from user's subsequent parameter manipulation.