Bug 1862.
EpetraExt does not depend on Triutils, but was failing when Triutils
was not built because some source files need to be conditionally
compiled.

I updated the --enable-epetraext-triutils option to conditionally
compile these files.

I tested the option with six different combinations of configure flags
and the option worked as expected every time.

Finally, I removed some obsolete Makefile.am's that were left over
from the days when multiple EpetraExt libraries were compiled (or at
least when that idea was being thrown around).