----------------------------------------------------------------------  Committing in .
 Modified Files:
 	Epetra_Import.h
 ----------------------------------------------------------------------
Bug 1570.

Some of the Doxygen documentation for the Import and Export classes
was lost a couple years ago.  It looks like the problem arose when
the list of inputs and outputs was added.  I am trying to reorder
some of the comments to see if I can get everything back on the
web pages.  If this is successful, I will make the same change for the
Export class.

Epetra still compiles on my machine.