I did a few things here:
1) I added and arcp_implicit_cast(...) templated conversion function.  Initially, I felt that no implicit conversions of ARCP objects should be directly supported but I realized that on imporant use case was not supported.  That is the conversion of ARCP<T*> to ARCP<const T*const> which is different than the conversion from ARCP<T*> to ARCP<T*const> which still allows the objects of type T to be modified.  However, the ARCP still does not define a templated constructors and therefore does not truely support an implicit conversion of ARCP<T1> to ARCP<T2> with will burn people from time to time.  I have added full testing for this as well as code in ifdefs that can be activated and show the correct compiler errors.

2) I changed the dim() function to be called the size() function which is consistent with the STL classes.

With these changes and additional tests, I am satisfied that all of the major features and type safty concerns have been addressed for this data type.  I still need to implement wrappers and unwrappers for Teuchos::Array objects but I will add those later or when they are needed (and that time will come).

Next I need to create a version of Teuchos::ConstNonconstObjectContainer called something like Teuchos::ConstNonconstArrayContainer that performs the same task of protecting const at runtime, except does it for an array of objects.