Address bug 2334. The new function Epetra_Util::Create_OneToOne_Map returns a new 1-to-1 map created from the possibly-non-1-to-1 input map.
This function specifically accepts and produces Epetra_Map objects.
I will add another overloading of the function that operates on
Epetra_BlockMap objects. (Although Epetra_BlockMap is a base-class of
Epetra_Map, I think it is necessary to have separate functions so that
the user can choose the type of the new map that is returned.)

Create_OneToOne_Map is static, so it is not necessary to create
an instance of the Epetra_Utils class before calling it.