adding back some files that I missed yesterday
NOTE: I messed up when I did the branch yesterday. branch BAKER_REFACTOR1
contains the code BEFORE the big refactor. the new stuff is in the main branch.
this is going to break codes that rely on Anasazi, unless those codes revert to
BAKER_REFACTOR1 or we update them to use the new framework.