added getRitzRes2Norms() function to Eigensolver; implemented it in BlockDavidson and LOBPCG
in BlockDavidson, modified setSize() routine to invalidate the state of the solver when blocksize
or numblocks is changed. previous version had a bug where changes could cause a misalignment to occur,
where the current basis dim would not be a multiple of blocksize. newer behavior is simpler to document
and maintain.

added an absolute values around maxLambda/minLambda in SVQBOrthoManager; this provably
does not alter the behavior of the current code, but future modifications could change
this, and it's "the right thing to do".