Bug 2430 was present in Tpetra as well.
The wrong set of arguments was being given to the
tac_arg_enable_can_use_package macro. This was causing a configure failure
on Atlantis and should have been causing configure failures elsewhere.

I tested the fix to make sure that the error message did not occur on my
Linux box.