Bug 2689
Fixing the Epetra::DebugTools object.  This was broke using certain configure options.  It has a dependency (derives from) on TestUtils and test utils is only built if the tests are enbaled.  Currently if the tests were disabled and epetra support enabled, the Epetra::DebugTools were still being built.  This would cause a link failure though.

Additionally, there was an include with a relative path in a file.  We NEVER use relative or absolute paths in any source code.  configure will set the include paths correctly. Do not do this:

#include "../test/utils/NOX_TestCompare.H"

Instead, always do this:

#include "NOX_TestCompare.H"

Also updated all examples that use the debugtools to include the debug tools
header directly.  It is no longer in the NOX_Epetra.H file since it has been
moved into the test/utils file.

All tests in serial and parallel are passing.