Several changes:
-- Restructured the traits classes to more closely follow template
metaprogramming styles and added a number of new traits classes.

-- Templated SFad and DFad on two types:  ValueT and ScalarT for when nesting
AD types for higher derivatives.  Using the new traits classes, ScalarT can
usually be deduced from ValueT, so it shouldn't be necessary to ever
explicitly specify it.  DFad and SFad have a specialization for when these
types are the same which gets rid of an extra constructor.

-- Modified the nested dfad example to use the extra constructor when nesting
AD types.

-- Modified the arguments the +=,-=,*=,/= operators for {S,D}Fad to only
take Expr<> classes instead of any type since the implementation of these
operators assumes an Expr interface.