Overhaul of the memory profiling function, ML_memory_check(), and supporting function ML_MaxAllocatableSize().  On Linux, it will now query the file /proc/meminfo.  On Red Storm, it will use heap_info().  Otherwise, it will use malloc() & mallinfo().  Added processor id's to max and min stats.  Added swap line.  Changed units to megabytes.  This is easier to read, and avoids overflow problems.
This is still in progress.  There still seem to be some problems on Red Storm,
but this could be due to that platform's special malloc library.

I did not update the memory query functions used by
ML_MultiLevelPreconditioner().  I'll open a bug to record that this still
needs to be done.