Bug 3039
Here I have added code to make it easier to define and work with validated
parameter lists.

Basically, all you do is to setup your valid parameter list with your
validators and documentation (very easy now), for example, in the
implementation of your getValidParameters() function.  You then simply call:

  paramList->validateParmetersAndSetDefaults(*this->getValidParameters(),dep);

then all of the parameters are validated and these not set are set to their
default values.  Also, any parameters that have a different numericl type than
is wanted are converted automatically.  You set the depth to validate and set
defaults with 'dep'.

For examples, see the updated code:

  teuchos/test/FancyOutputting/FancyOutputtting_test.cpp

  teuchos/test/ParameterList/cxx_main.cpp

  aztecoo/thyra/src/AztecOOParameters.cpp

for how to use these new features.

The class AlgorithmA from:

  teuchos/test/FancyOutputting/FancyOutputtting_test.cpp

is a very simple example of how to implement the Teuchos::ParameterListAcceptor
interface using these new features.

The code in:

  aztecoo/thyra/src/AztecOOParameters.cpp

shows how to use automatic number converters, and string-to-integral
validators in a simple way and only having to define the validators and the
default parameters values right in the call that sets the valid parameters.
The developer never acually sees the validators.

Very nice if I must say so :-)