Whew! Tough bug.  In Zoltan_PHG_Fill_Hypergraph, we were maintaining a count of the number of edges a vertex is in.  This could is used when ADD_OBJ_WEIGHT="pins", which means the vertex weight should equal the number of edges the vertex is in.
When Fill_Hypergraph is called from Zoltan_LB_Eval, all edges are
moved to the "removed" list.  Edges may also be moved to the "removed"
list when they are very dense.  The pins in removed edges were not
being counted in the ADD_OBJ_WEIGHT.  In other words, when called from
Zoltan_LB_Eval, if ADD_OBJ_WEIGHT was "pins" and that was the only
vertex weight, all vertex weights appeared to be zero.

Now we keep a second total of *all* edges a vertex is in, removed
or not removed, and use that total as the "pins" weight.