When hypergraph callbacks are defined, we add two lines to LB_Eval output giving cutl and cutn respectively: Zoltan_LB_Eval  Hyperedge (k-1)-connectivity cut:           21 Zoltan_LB_Eval  No. cut hyperedges:                         17
Now we also get these two lines when graph, but not hypergraph,
callbacks are defined.  The hyperedges are each vertex with all
it's neighbors, and the hyperedge weight is the sum of the weights
of the edges from the vertex to each of it's neighbors.

The label "No. cut hyperedges" is not very descriptive, since
the cuts are weighted by the edge weights.