SUMMARY:  Updated makesystem to deal with several issues BUGS:     No bug sugmitted COMMENTS: See Below
1)   Did you add any source files? No

2)   Did you update or change the build system? Yes
2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
     Makefile.am files and the Makefile.in files? Yes
2.b) [Optional but recommended] Did you blow away the libraries and
     executables and then rebuild the libraries? Yes

3)   Did you rerun the test suite on at least the packages that you
     changed? Yes

EXTENDED COMMENTS:

I make several changes to the build system here:

* Added TAC_ARG_INFORM_SUBPACKAGE_USED(...) macros to see if stratimikos
supports amesos, aztecoo, belos, ifpack, and/or ml.

* Fixed up Malefile.export.rythmos.in to have separate
RYTHMOS_TEST_[INCLUDES,LIBS] macros for the tests.  This keeps the
RYTHMOS_[INCLUDES,LIBS] macros clean with only what clients need to build and
link against Rythmos, not the tests or examples.

* Add link commands for the diagonalTransient.exe example for different linear
solvers and modifed the rythmos/test/definition file to use the linked names.
In this way, if a linear solver is not enabled, then the test will not even be
built.  This needs to be done for the other tests in rythmos.

* Added a diagonalTransient.exe test (DIAGONAL2b) using Belos just to use belos
for a test.  I discovered a bug in the
Thyra::BelosLinearOpWithSolve::describe(...) where an RCP would be
dereferences if unitialized andI fixed this with an extra boolean test.

* Added the class name to Rythymos::BackwardEulerStepper::describe(...) and
Rythymos::ImplicitBDFStepper::describe(...) when it is uninitialized.  This is
the only clue that I have in some verbosity modes to see what stepper is
actually being used.

* Epetra_CrsMatrix.cpp: I removed illegal (...) around a macro that expands to
braces and replaced it with braces.  This is not legal ISO C++ as shown by the
compiler message:

  Epetra_CrsMatrix.cpp:817: error: ISO C++ forbids braced-groups within expressions

However, the epetra test RowMatrixTrasposer_test.exe fails on my machine and I
submitted bug 3384 to Epetra.  The epetraext test Transpose_test.exe is also
failing.

Actually, I make the above change to Epetra_CsrMatrix.cpp but I am not
checking it in since these two tests are failing and I have not heard back
from Mike Heroux yet.