SUMMARY:  Changed from Teuchos::RefCountPtr to Teuchos::RCP BUGS:     None filed COMMENTS: See Below
1)   Did you add any source files? Yes
1.a) Did you do a 'cvs add'? Yes
1.b) Did you add them to a Makefile.am file in some way? [???]
1.c) Did you do a new bootstrap? [???]

2)   Did you update or change the build system? Yes
2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
     Makefile.am files and the Makefile.in files? [???]
2.b) [Optional but recommended] Did you blow away the libraries and
     executables and then rebuild the libraries? [???]

3)   Did you rerun the test suite on at least the packages that you
     changed? [???]

EXTENDED COMMENTS:

Here I have changed the name from Teuchos::RefCountPtr to Teuchos::RCP.

To allow for backward compatibility I have kept files named
Teuchos_RefCountPtr[Decl].hpp.  These files just include the new
Teuchos_RCP[Decl].hpp files and include the evil #define:

  #define RefCountPtr RCP

As long as client code explicitly includes the file Teuchos_RefCountPtr.hpp,
the should be no modification needed to absorb this refactoring.  However, I
had to modify a bunch of files in several Trilinos package to replace forward
declarations for Teuchos::RefCountPtr with the header Teuchos_RefCountPtr.hpp.
This should not be harmful to compilation time for *.cpp files since I can't
imagine a non-trivial *.cpp file that does not need the definition (or at least
the class declaration) of this class.

I have created the script

  Trilinos/packages/teuchos/refactoring/change-RefCountPtr-to-RCP-20070619.sh

that we automatically update all files in a package to change from RefCountPtr
to RCP and will leave out the Teuchos_RefCountPtr.hpp file containing the evil
#define RefCountPtr RCP.

I have run this script successfully on the packages didasko, moocho, rtop,
rythmos, stratimikos, teuchos, thyra, and tpetra.  For all of the other
packages, I left them alone unless they did not compile in which case I made
small adjustments (usually to add an include to Teuchos_RefCountPtr.hpp and/or
remove a forward declaration for RefCountPtr).

As part of this refactoring, I also changed the name from
Teuchos::ArrayRefCountPtr to Teuchos::ArrayRCP to be consistent and these
changed the file names to Teuchos_ArrayRCP[Decl].hpp.  Since no code outside
of Teuchos is using this class (yet) this should not be a big deal.

I build the serial and the MPI version.  There are a few failed tests in each
but these are known about for the most part and this type of refactoring
should not have any affect on runtime behavior once you compile and link
successfully.  I have not built the MPI version for awhile so I don't have a
local frame of reference to know if these tests have been failing or not
already.

Here are a few miscellaneous things I also did:
-----------------------------------------------

* Changed the name from Thyra::relErr(...) to Thyra::relVectorErr(...) in
Thyra_TestingTools[Decl].hpp.  This removed an overload that was confusing the
PGI compiler on reddish.

* Replaced forward declarations for Teuchos::RefCountPtr with an include for
Teuchos_RefCountPtr.hpp in a lot of files.  A forward declaration did not work
of course after the refactoring.  Again, you really don't gain anything with a
forward declaration for the RCP class since you need it's definition in almost
every *.cpp file.

* Added using Teuchos::RCP and some other using declarations to the Thyra
namespace in the file Thyra_OperatorVectorTypes.hpp.  I will do the same for
the Rythmos namespace soon.

* Did some source code reformatting in files in Thyra where I manually replaced
Teuchos::RCP with just RCP.