SUMMARY:  Ported thyrasoftware to IBM AIX machine sais028 and           SUN machine sass8000
BUGS:     2692, 2984

CHECKLIST:

1) Did you add any source files? No

2) Did you update or change the build system? No

3) Did you rerun the test suite on at least the packages that you
   changed? No, I did not run any tests ...

COMMENTS:

Here I have changed enough code in the thyra pacakge to get the libraries,
tests, and examples in the thyra package to compile and link.  I can not run
the tests since the mpiexec script complains about a hostfile problem or
something.

For IBM AIX machine sais028:
----------------------------

The main thing I did here was to change the names of several different
overloaded virtual functions to have different names and then I use the NVI
idiom to put in public non-virtual functions with the old names to keep
current clients compiling.  Specifically, where are the overloaded functions
that I renamed and replaced with NVI functions:

  MultiVectorBase::applyOp(...)
  =>
  mvMultiReductApplyOpImpl(...),
  mvSingleReductApplyOpImpl(...)

  MultiVectorBase::acquireDetachedView(...)
  =>
  acquire[Nonconst]DetachedMultiVectorViewImpl(...)

  MultiVectorBase::freeDetachedView(...)
  =>
  freeDetachedMultiVectorViewImpl(...)

  MultiVectorBase::commitDetachedView(...)
  =>
  commitNonconstDetachedMultiVectorViewImpl(...)

  VectorBase::acquireDetachedView(...)
  =>
  acquire[Nonconst]DetachedVectorViewImpl(...)

  VectorBase::freeDetachedView(...)
  =>
  freeDetachedVectorViewImpl(...)

  VectorBase::commitDetachedView(...)
  =>
  commitNonconstDetachedVectorViewImpl(...)

The above changes where enough to get all of thyra to compile and link.
However, there are still several more overloaded virtual functions that need
are just ripe to cause more problems and need to be changed also.  This
initial refactoring, however, proves that getting rid of overloaded virtual
functions fixes the problem.

I also had to put in some explicit Teuchos:: qualifications for describe(...)
even through I had 'using Teuchos::describe' at the top of each function.

These refactorings show that this compiler is not honoring 'using'
declarations correctly.  It seems to be honoring 'using' in most cases (I use
this all over the place) but it just gets confused in some cases.

There was also a problem with an anasazi makefile but I could not rebootstrap
since I don't have the right versions of autoconf or automake on this machine.

For SUN machine sass8000:
-------------------------

For this machine I had to do a few dumb things:

* The compiler could not figure out calls to Thyra::block2x2(...) when passing
in Teuchos:null.  To fix this, I just create a dummy null RCP<const LOB>
object called nullOp and passed that in.

The linker on this machine takes an amazingly long amount of time to work.

I need to remove some ifndef __sun disables to see if this code works with
this version of the SUN compilers.  These where put in at least a year ago to
get this working with paunchy.