Several changes for bug #3437:
1)  Got the LOCA::Thyra::Group going.  Right now we can only do
continuation, turning points, and pitchforks with bordering.  No bordered
solver yet, so no Householder/minimally augmented methods, and no complex
support.   Also there is no DerivUtils support yet so we can't get df/dp
from the ModelEvaluator (standard finite differencing as usual is supported).

2)  Added two Epetra tests that use the LOCA::Thyra interface.  These use
the Sacado FEApp to implement the tcubed problem.  The problem is built
into the LOCA test problems library and the LOCA_EpetraExt_ModelEvaluator
test was modified to use it instead of its own.

3)  Enabled Thyra multivector support and implemented Jacobian-multivector
operations in NOX::Thyra::Group.  Technically, the way LOCA uses multivector
views is incompatible with the implementation of views in Thyra.  However
LOCA only uses contiguous views, so according to Ross, it should work even
though the interface doesn't guarantee it will.  So far with tests added in
(2) they seem to work.

4)  Fixed some uninitialized values in the Moore-Spence groups which were
uncovered by the Thyra interface due to Ross' judicious use of NaNs in
unitialized vectors.

5)  Replaced brusselator.cpp with brusselator.C in the NOX/LOCA Sacado
FEApp examples to be consistent with the rest of NOX.