SUMMARY:  Fixing makall script so that if make fails in any loop for           any package then all of make will be exited.
BUGS:     3447, 3542

CHECKLIST:

1) Did you add any source files? No

2) Did you update or change the build system? Yes
2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
     Makefile.am files and the Makefile.in files? Not needed
2.b) [Optional but recommended] Did you blow away the libraries and
     executables and then rebuild the libraries? Not needed

3) Did you rerun the test suite on at least the packages that you
   changed? Not needed

COMMENTS:

This fixes a serious bug.  Previously, the makeall script's loop would
continue on even if a package make failed.  This was causing the test harness
to report build successes even through the build failed (see bug 3542).  This
version of the scripts for loop with exist on the first package build failure
and the return code from the makeall script will be the return code from the
failed make.

I tested this manually by breaking a test in teuchos and then making sure that
the top level 'make tests' command failed immediately after teuchos and would
not loop to any other packages.