SUMMARY:  Fixed RCP Node tracing
BUGS:     3577

CHECKLIST:

1) Did you add any source files? No

2) Did you update or change the build system? No

3) Did you rerun the test suite on at least the packages that you
   changed? Yes,  I have built and tested both with
   TEUCHOS_SHOW_ACTIVE_REFCOUNTPTR_NODES enabled and not enabled.

COMMENTS:

Here I have fixed a nasty little problem where the RCP node tracing capability
is enable (e.e. when you define the macro
TEUCHOS_SHOW_ACTIVE_REFCOUNTPTR_NODES) then a segfault would occur in some
programs.  This was due to a sloopy implementaion of the Item 47 "Ensure that
global objects are initialized before they're used" in "Effective C++".

I have now changed the implementation of Teuchos_RCP.cpp so that the list of
RCP nodes (an std::map actually) gets created and destroyed by the constructor
and destructor of PrintActiveRCPNodes manage this list.  Now this seems to
work and valgrind is happy.

Also, I changeded the files Teuchos_RCPDecl.hpp and Teuchos_RCP.hpp so that
the function Teuchos::print_active_RCP_nodes(...) is now listed in the main
Doxygen documentation tied to the Teuchos::RCP class.  Now, when TEUCHOS_DEBUG
is enabled you can turn on tracing by just hand modifying the Teuchos_RCP.cpp
file and manually defining the macro TEUCHOS_SHOW_ACTIVE_REFCOUNTPTR_NODES and
recompiling just *.cpp and relinking your executables.  This massively lowers
the overhead of turning on and off the expensive RCP node tracing when
debugging support is already enabled.  This comes at the expense of a little
O(1) function call overhead in the constructor and destructor when
TEUCHOS_DEBUG is defined.  When TEUCHOS_DEBUG is not defined (i.e. in an
optimized build of the code, which is the default), then there is no overhead
at all and of course RCP node tracing is disabled.

I have also fixed the ArrayRCP class to compile again when
TEUCHOS_SHOW_ACTIVE_REFCOUNTPTR_NODES is defined.