SUMMARY:  Removed circular RCP references in AztecOO/Thyra adapters
BUGS:     3579

CHECKLIST:

1) Did you add any source files? No

2) Did you update or change the build system? Yes
2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
     Makefile.am files and the Makefile.in files? [Done]
2.b) [Optional but recommended] Did you blow away the libraries and
     executables and then rebuild the libraries? Yes

3) Did you rerun the test suite on at least the packages that you
   changed? Yes

COMMENTS:

* Thyra_AztecOOLinearOpWithSolveFactory.cpp: I removed two calls to
set_extra_data(...) that where clearly creating circular RCP references with
Ifpack and ML preconditioners.  The stratimikos test that was showing the
undeleted RCPs now shows no remaining RCP objects!  I also cleaned up the
formatting of the file to conform to the TCDG document.  I will soon
clean up this file by breaking out a bunch of different subroutines and
enable more reuse.  It would also be nice to allow the reuse of internal
aztec preconditioners but this is not needed for the milestone.

* Teuchos_RCP.cpp: I added a call to std::cout so that the std::cerr output for
the RCP nodes would come last.

* I also changed the Makefile.am files for the stratimikos tests and examples to
make the real executable a dependency of the symbolic link targets.  This
fixes build errors when building in parallel where MAKE fires off the symbolic
link target before the main executable is built.