Merged isorropia changes made to the 8.0 branch back to the dev branch.
Most changes had already been merged back.  Now all changes have been
merged.

In addition, I incremented the version number for the dev branch to 2.1d.
If the developers prefer 3.0d because large changes are again going to be
made before Trilinos 9.0, the version can easily be changed again.  Also,
even if the version remains 2.1d, the version for the 9.0 branch can still
be 3.0 if appropriate.