run_zoltan_sparse_matrix now calls the Zoltan_MP_* query functions after load balancing.  (MP stands for "matrix partitioning".)  These queries return the partition and process assignment for rows, columns or non-zeroes.
zdrive should actually can an "LB_Eval" type of function, which would
call these queries to compute a balance measure.  But this is not written
yet.

This is tested on hg_simple.  I wanted to test as well on a non-symmetric
matrix, to ensure that I didn't confuse rows and columns somewhere.