Summary: Fixing intermittent hang in operator profiling in parallel.
Bug: 3646

Comments: In ML_Operator_Profile, an global MPI reduction operation was guarded
by an if that checked whether a timer was zero/nonzero. For a simple 3
processor run, one processor's timer was zero, while the other processors'
timers were nonzero, resulting in only two of the three processors
calling the global reduction.

The fix is to have the "if" rely on a global sum of the timer.

Also added apply-only time detail for the processors with the highest/lowest
comm+apply times, respectively.