SUMMARY: Adding Trilinos_version.h to give macros that announce the Trilinos version.
BUGS: 3680

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? Yes
1.a) Did you do a 'cvs add'? Yes
1.b) Did you add them to a Makefile.am file in some way (either as
     a header or soruce file or to EXTRA_DIST)? Yes
1.c) Did you do a new bootstrap? Yes

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No.  This header file
   is not currently included by any other header files in Trilinos so
   there is no reason to rebuild.

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? No

COMMENTS:

Here I have added a new header file Trilinos_version.h that defines the macros
TRILINOS_MAJOR_VERSION and TRILINOS_MAJOR_MINOR_VRESION.

This header file has no other dependences and is currently not included by any
other header files in Trilinos.  The purpose of this header file is to allow
APP code (e.g. Nevada, Alegra, Charon etc.) to automatically be able to query
what version of Trilinos they are building against.

This file gets installed in $prefix/include after installation.

I tested this by doing an install and then building the following moch
application, trilinos-vesion.cpp, to show that this works:


    #include <iostream>

    #include "Trilinos_version.h"

    int main() {

    #if TRILINOS_MAJOR_VERSION == 00
      std::cout << "\nTrilinos Dev\n";
    #elif TRILINOS_MAJOR_VERSION == 8
      std::cout << "\nTrilinos 8.0\n";
    #elif TRILINOS_MAJOR_VERSION == 7
      std::cout << "\nTrilinos 7.0\n";
    #else
      std::cout << "\nIllegal Trilinos version!\n";
    #endif

      return 0;

    }

I built and tested this as:

  $  g++ -I /home/rabartl/PROJECTS/install-dir/gcc-3.4.6/trilinos/dev/serial/dbg/include trilinos-version.cpp -o trilinos-version.dev.exe
  $ ./trilinos-version.dev.exe
  Trilinos Dev
  $  g++ -I /home/rabartl/PROJECTS/install-dir/gcc-3.4.6/trilinos/8.0/serial/dbg/include trilinos-version.cpp -o trilinos-version.8.exe
  $ ./trilinos-version.8.exe
  Trilinos 8.0

We will test this soon in Charon/Nevada and see if this needs to be tweaked or
not ...