SUMMARY: Fixing non-member swap(...)
BUGS: None filed but breaking build on tbird

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No, see below

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? No, I only retested teuchos

COMMENTS:

Here if fixed the delcaration for the non-member swap(...) which is included
to mirron std::vector but was not actually used or tested.  This caused a
compilation error on tbird (see email from Eric Phipps).

Here is the diff:

    [rabartl@gabriel Trilinos]$ cvs -q diff packages/teuchos/src/Teuchos_Array.hpp
    Index: packages/teuchos/src/Teuchos_Array.hpp
    ===================================================================
    RCS file: /space/CVS/Trilinos/packages/teuchos/src/Teuchos_Array.hpp,v
    retrieving revision 1.15
    diff -r1.15 Teuchos_Array.hpp
    76c76
    <   friend void swap( const Array<T> &a1, const Array<T> &a2 )
    ---
    >   friend void swap( Array<T> &a1, Array<T> &a2 )


I also added a test for the non-member swap in Array_test.cpp.  I actually
added the test first, confirmed that it did not compile and then fixed the
declaration as shown above.

I only rebuilt teuchos and retested teuchos even through this class is used
all through Trilinos an other codes.  However, given the nature of the change,
I think I am pretty safe in doing this.