Made a few changes to allow the tests to run correctly when the installation directory is not an immediate subdirectory of the source directory.
(1) Added an install-data-hook that copies mesh files and solver parameter files to the installation directory.

(2) Changed the path definitions in SundanceDefaultPath.hpp to look in the
source directory or the installation directory.  Note: in doing this, I disabled searches relative to the "Nihilo" directory. This is because Nihilo is going away. Run tests from the Sundance build/install directory and it will work.