SUMMARY: Ifdef-ing out dead code causing problems with SIERRA bjam build.
BUGS: None filed.

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (including adding
   files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? No.

4) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? No, I could not find anywhere
   in Trilinos where these files where being built!

COMMENTS:

The files EpetraExt_Dirichlet_CrsMatrix.[hpp,cpp] and
EpetraExt_Dirichlet_Vector.[hpp,cpp] appear to be dead as they are not
included in any build that I can find in Trilinos.  This was causing the bjam
build to fail.  Rather than have to modify the bjam files all of the time, I
would like to start a convention where dead source that is not removed gets
ifdefed out with:

    #if 0 // dead source

    ...

    #endif // dead source

I would also like us to find a way to handle other conditional capability
support through ifdefs and macro definitions instead of selective file
inclusion.

Here are the diffs:


    Index: src/transform/EpetraExt_Dirichlet_CrsMatrix.cpp
    ===================================================================
    RCS file: /space/CVS/Trilinos/packages/epetraext/src/transform/EpetraExt_Dirichlet_CrsMatrix.cpp,v
    retrieving revision 1.1
    diff -r1.1 EpetraExt_Dirichlet_CrsMatrix.cpp
    0a1
    > #if 0 // dead source
    100a102
    > #endif // if 0

    Index: src/transform/EpetraExt_Dirichlet_Vector.cpp
    ===================================================================
    RCS file: /space/CVS/Trilinos/packages/epetraext/src/transform/EpetraExt_Dirichlet_Vector.cpp,v
    retrieving revision 1.1
    diff -r1.1 EpetraExt_Dirichlet_Vector.cpp
    0a1
    > #if 0 // Dead source?
    54a56
    > #endif // if 0

With this, Aria/SIERRA + Trilinos Dev builds and all of the tests run!