Bug 2826. Removing the check for the BLAS and LAPACK. Very few packages (maybe just epetra and teuchos) should have to check for these libraries. I recompiled to verify that things work.
I also added belos to the list of optional dependencies in both configure.ac
and Makefile.export.didasko.in.  Previously the Belos examples tended to work
because the information would be obtained through the NOX Makefile.export file.
I happened to build with didasko and belos, but without NOX and found that
it didn't work.