SUMMARY: Basic refactoring of RTOpPack::[Const]Sub[Multi]VectorView classes to use ArrayRCP and not raw C++ pointers
BUGS: 3828

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? Yes
1.a) Did you do a 'cvs add'? Yes
1.b) Did you add them to a Makefile.am file in some way (either as
     a header or soruce file or to EXTRA_DIST)? Yes
1.c) Did you do a new bootstrap? Yes

2) Did you update or change the build system (other than just adding
   source files)? Yes
2.a) Did you remember to do a bootstrap and do a 'cvs add' on any
     Makefile.am files and Makefile.in files you created? Yes

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? Yes, I did
   complete rebuilds serial debug, serial optimized, and mpi debug.

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, all 824 serial and 677 MPI tests
   ran passed!

COMMENTS:

Here I have began the process of making RTOp memory safer.  This checkin
includes an update to the RTOpPack classes [Const]Sub[Multi]VectorView to
replace the raw C++ pointer for the array of values with a Teuchos::ArrayRCP
object.  I left deprecated version of the constructors and initialized
functions in place but I changed the values() functions to return the ArrayRCP
object instead of the raw C++ pointer.  This required changing lots of other
Trilinos code that was just using values() to now use values.get().  Later, I
will grep on values().get() to see where I need to refactor to replace the raw
pointers.

I also added very strong unit tests for the changed classes
RTOpPack::[Const]Sub[Multi]VectorView.  I have resolved to turn over a new
leaf and write comprehensive unit tests for all code that I am going to
refactor from now on.  The current literature from the software engineering
community is united on this.

I also refactored a good big of the RTOp software to use iterators into
ArrayRCP objects instead of raw pointers.  There is still a lot more work to
do to purge all raw C++ pointers for RTOp this this is a good start.

All of the other changes are to replace values() with values.get() where
needed.