SUMMARY:  Change test to take a argument for the input file
     CHECKLIST:

     1)   Did you add any source files?

     2)   Did you update or change the build system? Yes
     2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
          Makefile.am files and the Makefile.in files?  yes
     2.b) [Optional but recommended] Did you blow away the libraries and
          executables and then rebuild? no
     3) Did you rerun the test suite on at least the packages that you
        changed? yes

     COMMENTS:  test_simple failed last night, because during nightly
testing it doesn't actually run in the directory that it's built in.
It reads a matrix market file from that directory.  So I'm adding a
command line argument to specify where the input file is and chaning
the definition file to give test_simple that information.  I re-ran
runtests and it worked this time.