SUMMARY:   [Updating Sundance configuration & build system to work within Trilinos build tree]
     BUGS:      [3848]

     CHECKLIST:

     1)   Did you add any source files?  [No]

     2)   Did you update or change the build system?  [Yes
     2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
          Makefile.am files and the Makefile.in files?  [Yes]
     2.b) [Optional but recommended] Did you blow away the libraries and
          executables and then rebuild?  [Yes]

     3) Did you rerun the test suite on at least the packages that you
        changed?  [Yes]

     COMMENTS:

     [These changes make the C++ Sundance libraries work within a Trilinos build tree. PySundance hasn't yet been tested (and probably doesn't work) and
 it's not yet possible to build Sundance against an external Trilinos. Those
 changes coming soon.

More detail on the changes:
(1) The trilinos version used by Sundance now defaults to TRILINOS_DEV.
(2) A flag --withtrilinosdatadir has been added, allowing the user to
specify the location of the mesh files for certain Sundance tests.
(3) The --with-trilinos flag defaults to $pwd/../.., so that by default
a Sundance build is assumed to be living within a Trilinos build.
(4) Changed the default search path for meshes and xml files to reflect the
changed location of the build.
(5) Fixed a problem in the ExodusMeshReader source, whereby the mesh search path
wasn't being used.
(6) Changed the interpolation order in Poisson3D.cpp to 2nd order

]

(checklist version 2007-06-20)