SUMMARY:   Removed MPI operator that was incompatible with Open MPI.
     BUGS:      NA

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

HSFC used an MPI operator that was incompatible with Open MPI.
The defined MPI operator looped over an array of size len, performing
summation on the first len/3 entries, max on the next len/3 entries, and
min on the last len/3 entries.  This works as long as all processors
receive the full array of size len (as they do in LAM).
But in OpenMPI, each processor performs work on a subset of the array, so
the value of len is not what the function expects.  As a results, the
operation is not performed correctly.

I replaced the MPI operator with three calls to MPI_Allreduce, one each for
MPI_SUM, MPI_MAX, and MPI_MIN.

This problem was exposed on a CEA dataset (conesphere1m) on Tbird, but
was reproducible on octopi with OpenMPI.

We've fixed this problem once in PHG; perhaps we should examine all our
MPI operators for correctness.

(checklist version 2007-06-20)