SUMMARY: Some more RTOp refactoring and fixing std::list usage problem
BUGS: None filed

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? Yes
1.a) Did you do a 'cvs add'? Yes
1.b) Did you add them to a Makefile.am file in some way (either as
     a header or soruce file or to EXTRA_DIST)? Yes
1.c) Did you do a new bootstrap? Yes

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? Yes, both serial debug
   and opt optimized

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, all 880 serial debug and 710 mpi
   optimized tests run passed.

COMMENTS:

(*) Teuchos_ConfigDefs.hpp: Added ifdef logic not to define _GLIBCXX_DEBUG if
it is already defined.  This avoids some annoying warnings.

(*) MoochoPack_LineSearchFilter_Step.[hpp,cpp]: Changed code that removes
entries from the filter to use the built-in std::list member function
remove_if(...) and defined a predicate class/object to determine if the entry
is to be removed or not.  This gets around some undefined code that was
written by a student years ago that the checked STL with g++ caught.  Now the
checked STL is 100% happy with MOOCHO!

(*) rtop: Added unit tests for ROpMax and ROpMin, refactored to use macros to
create the classes and expended support for custom-defined intial reduction
values to support this.  I am getting closer to where I will have the
RTOpPack::RTOpT interface completely cleaned up to not include any raw C++
pointers.  One step at a time ...