SUMMARY:  Change to try to address compiler bug on s858352
BUGS: 4010

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? No, I only changed one simple file

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No, just rtop with serial
   debug

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, but only rtop in serial debug
   mode with 19 passing tests

COMMENTS:

This is an attempt to deal with the compiler bug on the MAC machine s858352
(see bug 4010).  I would just directly log onto the machine but my current
password does not work.  I am not sure who maintains this machine so that I
can get a new password but I will make this one lucky try to fix the issue.