Subject:  Adding modified SuiteSparse 3.0 library to Amesos
BUG #:  3996

Summary:

This check-in is the first step in addressing the separation of KLU / Paraklete
solvers from the SuiteSparse library to avoid potential issues with supporting 3PLs
that also depend on the SuiteSparse library.

The SuiteSparse v3.0 library has been modified to change the filenames and
function names to be prefixed by "amesos_".  This means the headers that Trilinos
installs and the symbol table from the Amesos library will not include standard
headers and symbols that are found in the SuiteSparse library (AMD, BTF, CHOLMOD, etc.).

Furthermore, the modules of CHOLMOD that are GPL licensed are explicitly
removed from the SuiteSparse v3.0 library to avoid problems for users of Amesos/Trilinos.
These modules are not essential for functionality and CHOLMOD provides work-arounds
that are already in use by Amesos.  It's just that the source code was never removed.