SUMMARY: Added waitAll(...) and test and getRawPtr() to all MM classes
BUGS: 3828, 4011

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes, I updated faily recently but I did not rebuild every
   thing since then

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No, I just
   rebuilt all of Trilinos, I did not 'make partial-clean'.  I did
   do serial debug and mpi optimized modes through.

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, all 924 serial debug and 646 mpi
   optimized tests run passed!

COMMENTS:

Here Chris Baker and I added implementation for Teuchos::MpiComm::waitAll(...)
and unit test.  We did not implement any of the non-blocking functions for
Teuchos::SerialComm.  The class should implement non-blocking functions be we
just did not want to take the time to do this.  I think this Lean software
development at it's best :-)

The unit tests for non-blocking simply return success==true when using a
Teuchos::SerialComm (determined with a dynamic cast).

As part of this checkin, I also added getRawPtr() to all of the memory
management classes.  This is to make a consistent way to get at the raw
pointer for all of these classes.