SUMMARY: Partial port of Thyra to Sun 10 sass8000
BUGS: 3978

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No, I only rebuilt
   a few files on the Sun

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? No, it is not easy to set
   up to test this because of how everything is layed out to build
   locally on sass8000.

COMMENTS:

The Sun compiler found a defect where I was not overriding a set of virtual
functions in Thyra::VectorDefaultBase for Thyra::MultiVectorBase the way I
throught I was.  I obviously did not test these functions anywhere so this is
not good.  I also added 'virtual' to all of the virtual functions being
overridden (as suggested in "C++ Coding Standards").

There is a lot more work to do to port Thyra to the Sun so that it will also
work with other platforms like purple.sandia.gov.  I will work mostly on my
linux box gabriel.sandia.gov and then come back to the Sun.