SUMMARY: Ifdefing out using declarations needed by the sun 5.7 compiler
BUGS: 3978

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes, I think

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No, I just rebuilt
   thyra in serial debug mode

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? No, I just rebuilt

COMMENTS:

This is a bit of an emergancy checkin due to the fact that adding all of the
using declarations broke thyra on g++ compilers.  To try to fix this, I have
ifdefed all of the using declarations with THYRA_INJECT_USING_DECLARATIONS
which I do not have defined yet.  I will turn around and define this on the
sun and try it there.