SUMMARY:   improvements to Distributor and its unit testing
     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

Added comments to some particularly obfuscated bits of code in Tpetra::Distributor. Fixed some bugs.
Expanded unit tests of this class. createFromRecvs() test still failing (disabled temporarily), but
this may be a bug in the test and not a bug in the functionality. doPostsAndWaits() and createFromSends()
look pretty good.

(checklist version 2007-06-20)