SUMMARY:   Modification of BLAS explicit instantiation of templates to support native BLAS libraries
     BUGS:      4049

     CHECKLIST:

     1)   Did you add any source files?  No.

     2)   Did you update or change the build system? No.

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes

     COMMENTS:

     The BLAS support of native, optimized BLAS libraries used a partial template
specialization allowing the OrdinalType to be something other than int.  However,
the underlying BLAS wrappers assume the OrdinalType to be int.  This inconsistency
causes a compiler error when you try to use BLAS<long, double> or some instantiation
of the template for an OrdinalType other than int.  This has been fixed so that
explicit template instantiations are used with the BLAS wrappers.  Anything else
will use the templated BLAS reference library that lies within Teuchos.