Summary: Testing timeouts, how to set them, and see if they kill MPI processes
Here the only thing that I am really checking in is a time limit of 120
seconds (or two minutes) per test.  I played with this a little and I verified
that it does indeed kill all of the processes of a parallel run when it times
out.

NOTE: The ctest command-line argument --test-timeout does *not* appear to
work.  Instead, you need to use the CMake cache variable DART_TESTING_TIMEOUT
shown in the scripts that I am checking in.

Timeouts will save a lot of time and trouble and will help us to control the
development process by not letting people add tests that take too long.