SUMMARY: Fixing link error on purple.sandia.gov
BUGS: 4204

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? Yes, but I only
   built aztecoo with fortran support turned off.

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, but only the aztecoo test suite.

COMMENTS:

Here, I have fixed a link error that occurs on purple.sandia.gov with SIERRA +
Trilinos 9.0.  The problem was that the file az_c_util.c was calling BLAS
routines without using the portability macro F77_BLAS_MANGLE(...).

The reason that this has not bit other projects I suspect is that the SIERRA
BJAM Jamfile uses a glob to bring in all *.c files and therefore compiles this
file.

See the bug 4204 for all of the ugly details.

To test this, it build trough the SIERRA BJAM build system on
purple.sandia.gov but the code was not run.

I also build Aria + Trilinos 9.0 with this change on linux and the 'fast' test
suite ran just fine.

I also build the native Trilinos aztecoo tests suite with --disable-fortran
and it linked and all of the tests passed.  I verified that the fortran object
files az_reorder.o and az_f_util.o were not built but the files az_c_reorder.o
and az_c_util.c were.

I am pretty confident that this change is the correct thing to do, even if it
really should not impact SIERRA since SIERRA should not be building these
files in the first place.  However, I tried to exclude the files az_c_util.c
and az_c_reorder. and I got a link error (see bug 4204).  Therefore, I am
leaving well enough alone!