SUMMARY:   further development in Tpetra, Anasazi
     CHECKLIST:

     1)   Did you add any source files?  yes
     1.a) Did you do a 'cvs add'?  yes
     1.b) Did you add them to a Makefile.am file in some way?  yes
     1.c) Did you do a new bootstrap?  yes, both packages

     2)   Did you update or change the build system?  yes
     2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
          Makefile.am files and the Makefile.in files?  yes
     2.b) [Optional but recommended] Did you blow away the libraries and
          executables and then rebuild?  yes

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes. ran two builds, one with just tpetra, one with just anasazi. both build and install.

     COMMENTS:

Development of MultiVector and CrsMatrix is almost complete. Sufficient functionality exists to use for solving eigenvalue
problems or linear systems of equations. Additional unit testing is in place for these class; significantly more is needed
for an appropriate coverage of functionality. A few class methods are not filled in yet; I will continue to work on these.

Finished Anasazi wrappers for Tpetra. Authored two Anasazi tests using Tpetra: MVOPTester and a BlockKrylovSchur eigensolver test.
The eigensolver does not work, implying some bug(s) in Anasazi wrappers or Tpetra. I suspect it is in Tpetra::CrsMatrix, the less
well-tested component. Anasazi can be tested with Tpetra by specifying --enable-anasazi-tpetra (default: false). This is not supported
at this time.

I will immediately checkout and test on a different platform to ensure no errors were introduced in the build system because of the
introduction of new files.

(checklist version 2007-06-20)