SUMMARY:   fixed bugs, added tests
     CHECKLIST:

     1)   Did you add any source files?  no (added some test files)

     2)   Did you update or change the build system?  yes, added support for triutils
     2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
          Makefile.am files and the Makefile.in files?  yes
     2.b) [Optional but recommended] Did you blow away the libraries and
          executables and then rebuild?  yes, serial and mpi

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes, serial and mpi.

     COMMENTS:

Fixed numerous bugs in Tpetra::Distributor, Tpetra::DistObject, Tpetra::CrsMatrix.
No bugs are currently identified and all tests pass. However, even a simple power method (serial or MPI) fails to
converge on an eigenvector, and more complicated eigensolves (with Anasazi) fail as well. This is the focus for the
current debugging.

(checklist version 2007-06-20)