Summary: Finished automated intrapackage dependency handling
I have now finished the intrapacakge dependency handling capability.

Here is what I did:

(*) Added option Trilinos_ENABLE_ALL_FORWARD_DEP_PACKAGES and when it is ON,
all forward packages that depend on the set of enabled packages will be
recursively turned on.

(*) Added option Trilinos_ENABLE_ALL_OPTIONAL_PACKAGES and when it is ON, all
packages that are optional dependencies of the given set of packages is turned
on.

(*) Added support for automatically enabling all required packages for current
set of enabled packages.

I have tested all of this manually and it looks very good.  I still need to
add unit tests for dependency handling and refactor the code to clean it up a
bit.

I tested this with serial debug and mpi optimized