Summary: Cmakifying Thyra and porting to MS Visual C++ 9.0
Here, I have added the CMake build system for Thyra.  I wanted to
CMakify thyra before the TUG because Thyra is a little more complex
than the other packages done so far as Thyra as adapters for Epetra
and EpetraExt and build libraries and examples in other directoires.
Sure enough, these was a little change that was needed to get the
adapter libraries to build.

The key to the current CMake build system is that you have to include
the directoires containing libraries in the order that they depend on
each other.  For example, that is why the CMakelists files for the
libraries are processed in the following order:

  thyra/src
  thyra/adapters/epetra/src
  thyra/adapters/epetraext/src

Also, the tests and examples have to come after the libraries of
course.

NOTE: I could not add the epetraext example since it has a cicular
dependency that relies on stratimikos.  In the current CMake system,
we can not accommodate such circular dependencies.  What I need to do
is to make this example only rely on the
Thyra::DefaultSerialDenseLinearOpWithSolve[Factory] classes.  I should
also move the current version of this example to didasko but I will do
that once all of Trilinos is Cmakified.

I general, I have tried to fix up the include directories and link
library directories to maintain the correct ordering.  This is not
maintaining the ordering 100 percent correctly but it is better than
it was.  This should not be necessary if we name our source files and
libraries correctly and are not sloppy but you can never be too
careful.

Here are some of the specific things that I did:

(*) Trilnos/CMakeLists.txt:

- Set CMAKE_INCLUDE_DIRECTORIES_BEFORE=ON so that by default include
search directories get prepended instead of postpended.  This should
not be necessary but it does make the compile lines easier to read.

(*) TrilinosPackageLibraryHelpers.cmake

- TRILINOS_PACKAGE_ADD_LIBRARY(...):
-- Added INCLUDE_DIRECTORIES(${${PACKAGE_NAME}_INCLUDE_DIRS})
when this file gets included so that the headers that have been added
so far.
-- Added VERBOSE_CONFIGURE output

(*) Thyra_LinearOpBaseDecl.hpp: Had to replace
ScalarTraits<Scalar>::one() with static_cast<Scalar>(1.0) because the
MS Visual C++ compiler would not allow the former.  This is
unfortunate but I suspect that this change will be just fine.  I think
that we should expect any scalar type to be constructable from double
floating point numbers and conversions from 0.0 and 1.0 should make
some sense.

(*) Teuchos_FancyOStream.hpp: Removed the template arguments from
fancyOStream(...) and getFancyOStream(...).  This allows the MS Visual
C++ 9.0 compiler to make some function calls that it could not
otherwise.  I think we should just remove all of templating of the
FancyOStream class and just assume char.  This will improve usability
and portability.

(*) cygwin-gcc-3.4.4-serial-debug-cmake: Added
-DTrilinos_ENABLE_NATIVE_TEST_HARNESS:BOOL=ON to enable the native
test harness on cygwin.

I tested this just with cygwin/g++ and MS Visual C++ 9.0 on my MS
Windows laptop.

There are now 107 serial package tests run on cygwin.  Of these, the
only test that fails is:

    4 - Teuchos_Containers_test (SEGFAULT)

This has been reported as a bug and need to be removed from Teuchos
and moved to Sundance.

The following tests are failing with MS Visual C++ 9.0:

1>The following tests FAILED:
1>  80 - Triutils_read_test (Failed)
1>  83 - EpetraExt_inout_test (Failed)
1>  87 - Thyra_test_scalar_product (Failed)
1>  88 - Thyra_test_composite_linear_ops_serial (Failed)
1>  89 - Thyra_test_composite_linear_ops_spmd (Failed)
1>  94 - Thyra_test_linear_op_with_solve_n1_n2_0 (Failed)
1>  95 - Thyra_test_linear_op_with_solve_n1_n2_1 (Failed)
1>  96 - Thyra_test_linear_op_with_solve_n4 (Failed)
1>  97 - Thyra_sillyCgSolve_serial_0 (Failed)
1>  98 - Thyra_sillyCgSolve_serial_1 (Failed)
1>  99 - Thyra_sillyCgSolve_serial_2 (Failed)
1> 100 - Thyra_sillyCgSolve_mpi (Failed)
1> 101 - Thyra_sillyPowerMethod_serial (Failed)
1> 102 - Thyra_exampleImplicitlyComposedLinearOperators_0 (Failed)
1> 103 - Thyra_exampleImplicitlyComposedLinearOperators_1 (Failed)
1> 104 - Thyra_silly1DFFT_serial (Failed)
1>Errors while running CTest

I have not looked through these yet but given this is the first port
to MS Visual C++ 9.0, I don't think I broke the port :-)