SUMMARY:   added new test to harness, documenting LAPACK problems in Anasazi on QEVP problem from Salinas
     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes. only the new test fails.

     COMMENTS:

Added outputting of significant multivectors, serial dense matrices to BlockKrylovSchur. This outputting only occurs on
Debug level verbosity; however, I left this code commented out---look for Debug in iterate()

Modified BlockKrylovSchurQEVP_test to set initial multivector to all ones. This allowed me to run near identical problems, side-by-side in Salinas and this test.
Discovered parameter settings for this which cause failure, due to info != 0 error in LAPACK DGEES. Added this as a test to the harness.

(checklist version 2007-06-20)