Summary: Partial CMakeification of Isorropia
Here I have done just the most basic CMakeification of Issoropia.  I
only did a few tests and examples just to show how to do them.  Adding
the rest of the cases should be very easy for anyone to do given what
is there.

One issue that is unresolved is that Issoropia creates called
isorropia_test_utils that is only used for tests and examples.  Just
do this fast, I made this an offical issoropia library but we need to
upgrade the CMake macros to provide special handling for test
libraries.  They should not be installed and they hould not be added
to Issoropia_LIBRARIES.  This will be easy to fix later so I am moving
on.

I tested this by building from scratch with both cmake serial debug
and mpi opt and all tests passed.

Other minor things done:

(*) packages/teuchos/README.RELEASE_NOTES: Forgot to add this before.

(*) sampleScripts/linux-godel-gcc-4.1.2-serial-debug-cmake: Added
linker flag for -lm since zoltan now needs this.  I guess that the C
compiler does not bring this library in but the C++ compiler does by
default.

(*) cmake/TODO: Added item for dealing with test-only libraries