Summary: Cleaning out Amesos_config.h
Bugs: None filed

(*) amesos/cmake/Amesos_config.h.in: Cleaned out a bunch of junk that is not
being used for anything useful that I can see.  NOTE: This only affects
CMake-built code so I don't need to test the autotools build.

(*) Amesos_ConfigDefs.h: Removed the ifdef for <vector> include.  The same
behavior is maintained in that if you don't have <vector> then Amesos will not
compile.

(*) packages/fei/test/CMakeLists.txt: Changed number of processors to match
what was in fei/test/defintion so that it would pass.  I also added other test
cases but they timed-out or failed so I commented them out.

(*) cmake/package_arch/PackageAddTest.cmake: Added ELSEIF case for the a,b,c
form of NUM_MPI_PROCS and put error message in it since it is not supported
yet.  The FEI tests I played with but disabled used "NUM_MPI_PROCS 1,2,4" so
we will need to support this.  Without this, it just assumes the number of
proces is '1,2,4' which mpiexec is not going to like :-)

Builds/tests that I ran:

- CMake serial debug rebuild (All 235 tests ran passed)

- CMake mpi release rebuild (All but one of the 208 tests ran passed, see
below)

The test Tpetra_CrsMatrix_UnitTests_MPI_4 fails on gabriel.sandia.gov and it
has never been fixed.  I talked with Chris Baker about this.