Force all modification of vertex coordinates to go through functions in PatchData by removing the variants of PD::get_vertex_array() and PD::vertex_by_index() that return a non-const pointer or refernece, respectively.  Thus other code cannot modify the vertex coordinates through a reference to the MsqVertex instance.  The primary motivation for this is to allow PatchData to keep track of when vertices are moved so it can update other data (e.g. slave vertex coordinates.)  This also would have avoided some past bugs that were worked around in another way in some of the relaxation smoothers.  It also exposed a bug where a statement like "vtx = new_pos;", with vtx as a reference to a MsqVertex and new_pos as a Vector3D, would clear (zero) the bit flags on the MsqVertex.
Most of the modified files are simply the addition of the 'const'
keyword to a few lines.  Others are minor changes for setting vertex
coordinates.  The most complex changes (other than PatchData itself)
were to QualityMetric.cpp (numerical gradient/Hessian calculation),
some of the unit tests that modify the contents of PatchData in atypical
ways, and NonSmoothSteepestDescent (because touching anything in that
code is a lot of work.)