SUMMARY:   I found a minor bug as I reworked the Rythmos StepperBuilder      class to use the Teuchos ObjectBuilder.
     BUGS:      4347

     CHECKLIST:

     1)   Did you add any source files?  NO

     2)   Did you update or change the build system?  NO

     3) Did you rerun the test suite on at least the packages that you
        changed?  YES

     COMMENTS:

I found that if you added a bunch of object factories to the object builder and
then immediately asked for an object by name, it would end up in a place where
it did not have a valid parameterlist and it tried to dereference a null
pointer.  The fix was to be a little more aggressive about checking for a valid
parameter list and create one a little earlier.

(checklist version 2007-06-20)