SUMMARY:   Fixed most calls to copyMat that used bools instead of a scaling variable
     BUGS:      None.

     CHECKLIST:

     1)   Did you add any source files?  No.

     2)   Did you update or change the build system?  No.

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes.

     COMMENTS:

     Bill Bohnhoff observed that several of the calls to copyMat passed in a
boolean for the last argument instead of a scaling variable of scalar type.
The private copyMat() method has been around for some time and was changed
a while ago to take a scaling argument instead of an argument.  Unfortunately,
several uses of this private method were not updated.  However, the usage of
"false" and its conversion to "0" happened to result in the correct behavior.

(checklist version 2007-06-20)