Summary: Fixed an issue with make install-everything that prevent some files  from being copied into the install directory
Bugs: 4313

I reproduced the bug, and then fixed the bug, and then reinstalled, showing
that the bug was in fact fixed.

The problem was that make install was never run in the top level directory
during a make install-everything.  I added a little make target at the
package level to call at the end of make install-everything that is the
same as the make install hook at the top level.