Hack to allow multiple copies of the Mesquite library with different major version numbers to be linked into the same application.  The primary motivator for this is ddriv, which needs to link both Mesquite 2.0 and libcubiti19.so, where libcubiti19.so needs Mesquite 1.1.3.
Replace the "Mesquite" namespace with the "MESQUITE_NS" macro, which expandes
to Mesquite##MSQ_VERSION_MAJOR (in this case, "Mesquite2").  Declare
"namespace Mesquite" to be an alias for "namespace Mesquite2".  The end result
of this is that there is no visible change to the API for applications, but
the mangled symbol names in the library are different for different major
versions of Mesquite.