SUMMARY:   error in macro causing build errors (and test failure) on some platforms
     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  yes
     2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
          Makefile.am files and the Makefile.in files?  yes
     2.b) [Optional but recommended] Did you blow away the libraries and
          executables and then rebuild?  yes

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

Macro for test in CMake and Autotools that complex BLAS problem could be correct
duplicated a variable name. On some platforms, most notably the GCC 4.3.3 on s909348,
this caused the test to fail to compile, which caused the test to report failure, so
that the complex BLAS fix was not activated, causing test failures in Tpetra.

I correct the macro, tested under GCC 4.3.3 and 4.0.1 on s909348, test now detects
properly and failing Tpetra tests now pass.

(checklist version 2007-06-20)