SUMMARY:   DefaultObject feature
     BUGS:      4347

     CHECKLIST:

     1)   Did you add any source files?  NO

     2)   Did you update or change the build system?  NO

     3) Did you rerun the test suite on at least the packages that you
        changed?  YES

     COMMENTS:

I added a feature to set the default object.  This changes the valid parameter
list and only takes effect if the desired object is not selected either through
the parameter list or through the create() function.  This is useful when
you're creating a builder and you want to specify what the default is so its
valid parameter list has the right set of defaults downstream.



(checklist version 2007-06-20)