SUMMARY:   Stepper Validator (bare bones), plus debugging on the Appender
     BUGS:

     CHECKLIST:

     1)   Did you add any source files?   YES
     1.a) Did you do a 'cvs add'?  YES
     1.b) Did you add them to a Makefile.am file in some way?  YES
     1.c) Did you do a new bootstrap?  YES

     2)   Did you update or change the build system?  NO

     3) Did you rerun the test suite on at least the packages that you
        changed?  YES

     COMMENTS:

I started a StepperValidator class which will incorporate implicit requirements
on the steppers.  Currently it verifies that model parameters are passed
through the stepper to the model upon evalModel.  And it currently only works
for explicit methods and only has a unit test for ERK.

The PointwiseInterpolationBufferAppender had no tests, so I started some and
immediately ran into some problems.  I've added lots of various tests and debug
output to figure it out, but to no avail so far.  This is an intermediate
commit.

(checklist version 2007-06-20)