Comment back in unistd.h to fix Unix/Linux builds
Bugs: 4400

When doing the MS Windows port, I stupidly forgot to comment back in the
unistd.h file that I had commented out to test.  All of the builds with all
versions of GCC have been passing for many days after I checked in this
change.  This tells me that either the Trilinos tests are set up to already
include the unistd.h file from some other Trilinos include file or that g++ is
including this file in other header files were it should not be (a common
problem with g++).

Anyway, I have not tested this on Sun but given the nature of the changes to
this file, I can almost 100% guarantee that this will fix the Sun build error
that Alegra + Trilinos Dev is seeing.

I also fix a bunch if illegal conversions from 'const char*' to 'char*' that
g++ was complaining about.

I did test this on Linux/gcc in both serial debug and mpi release mode by
enabling only Ifpack and all of its dependent packages..