Updated Sacado and FEApp to work with latest Stokhos development.
FEApp now has a reasonable implementation of intrusive stochastic Galerkin
using Sacado, and element-level quadrature from Stokhos.

There is an experimental directory that I added that has some Dakota code
that I experimented with.  Turns out that wasn't the way to go, but I didn't
want to lose that code completely.  That code is not compiled and is just
here for posterity.

FEApp still needs an SG-aware model evaluator to simplify the SG implementation.
Here we are just creating a new nonlinear problem for the SG problem.