Fixed an issue with the fortran mangling configure test that was uncovered during the redstorm port effort
The fortran mangling configure test wasn't compiling properly do to needing
extra libraries on the link line that were being passed in via
Trilinos_EXTRA_LINK_FLAGS. The configure test wasn't paying attention to these
so I added them to the C_FLAGS and F_FLAGS. I had to do that as the cmake way of
doing this, which is to add them to LINK_LIBRARIES, didn't work. It seemed this
was being ignored. Not sure if that is a cmake bug or something with the way I
was using it though.

This change probably should be made to all the configure time try_compile
checks, but so far this is the only one that has bitten us and there could be
cases where adding the extra flags is the wrong thing to do too so I left them
alone until they can be looked at individually to see if this change makes sense
for them.