Initial checkin of checkin testing tool
This is an intial checkin of the new checkin-test.py script that will be used
for automated and standardized pre-checkin testing.  This script works right
now but you currently have to explicitly list the packages to enable.

I will add the auto package enable detection next and implement a few other
features.