Partial port to MS Visual C++ 2008 Express
Here I have done some work getting Teuchos, Thyra, Stratimikos and
MOOCHO to build and run on MS Windows with the MS Visual C++ 2008
Express compilier.  I don't have everything working but most of the
issues are being resolved.

Teuchos:
--------

All of the Primary Stable code in Teuchos should be bulding and all of
the tests run should pass.

(*) teuchos/test/CMakeLists.txt: Disabled the CTestScript tests since
they all rely on diff.  I could fix these later to use 'cmake -E
compare_files' but this is not a big deal for now.

(*) teuchos/test/TabularOutputter/CMakeLists.txt: Disabled the
outputter test since MS Visual C++ formates numbers different that GCC
or the Intel compilers.

Thyra:
------

I did not do anything for Thyra but there are several failing tests.
The problem seems to be that Thyra::randomize(...) seems to be
producing NaNs.  I will have to debug this later.

Stratimikos:
------------

I tried playing with tolerances but in the end, there was one test
with Belos based on a block RHS GMRES algorithm that just failed very
badly.  In the end, I just disabled one of the tests because I just
could not get it to pass.  This seems to suggest to me that their
might be something wrong with Belos or perhaps the BLAS on this
machine.  There was another set of tests that I just had to loosen the
tolerances a little.

MOOCHO:
-------

I had to fix some silly using declaration errors (they were valid C++
but MS Visual C++ did not see it that way) and now all of MOOCHO
compiles (withour Fortran).  However, almost every test in MOOCHO
fails.  The AbstractLinAlPack::VectorSpaceTester fails a test with a
norm_1 right away.  I will need some time to figure out how to debug
this but at least MOOCHO builds now.