In preparation of taking over the nightly testing on godel making some changes to the driver script to not refer to rabartl.
Also removed the reliance on having a .bash_profile, instead the important
things will be setup in the crontab. This will be a minimal set of env vars to
prevent my environment from polluting the testing.

Disabled the memcheck testing since it has only been hanging recently. Instead
of running memcheck tests on sunday for now we will run the regular nightly
testing. Coverage testing is unaffected by this.

Since where this testing lives has been changed to the trilinos "common" area on
godel I put in chgrp and chmod commands to change the group to trilinos-dev and
make everything read/writable by the group