remove the experimental marker from the serial tests on trilinos-test
the ld issue seems to have cleared up with the new version of cmake so I am
moving the serial debug build back to a nightly test so that people will get
notices if tests break. If the ld issue comes back again I will move it to
experimental again, but much sooner than I did this time.