Upgraded automatic documentation to convert Unicode to ASCII
Script doxy2swig.py now converts doxygen documentation that has
Unicode into a reasonable semblence in ASCII (e.g., and "a" with an
umlaut gets converted to an "a").  This fixes the Komplex_dox.i file,
which no longer contains Unicode, and so the resulting Komplex.py
module now imports to python properly.