Fixed misaligned pointer error in various test scripts
These test scripts were suffering from the misaligned pointer warning
that has been tracked to the use of module optparse.  In these cases,
the warnings were not causing a crash or failure, so ctest was not
interpreting the results as failures.

The fix, as before, was to get rid of the optparse module and replace
it with the getopt module.