convert testing to raw git from eg
eg doesn't work terribly well on windows. This is a product of its and git's
reliance on the standard gnu like environment. Git got around this by including
the required parts in the windows executable, eg being a perl script doesn't
have that option. Instead of hacking an environment together on windows that may
or may not be stable I felt it was better to just use raw git in our testing.
This does remove the testing of eg on the testing machines that we were getting