Added deprecated warnings support for RTOp and Thyra and refactored some code.
RTOp:
-----

Here I added support for RTOP_DEPRECATED and applied it to all RTOp functions
that were commented as deprecated.  I had to change a bunch of code because
RTOp is marked as CLEANED and therefore can not tolerate any warnings.

Thyra:
------

I also added support for THYRA_DEPRECATED and applied it to a few functions
(not much yet).  Mostly, I just cleaned up a bunch of warnings caused by
RTOP_DEPRECATED in Thyra.  This resulted in a lot of refactored code including
some code that was using raw C++ pointers.  This is a little scary given the
amount of code I changed.  I also converted some Thyra files over to use
explicit instantiation and fixed a bunch of missing references that always
occurs due to this.  There are still a lot of warnings left but you have to
checkpoint now and then and integrate.

A major thing I did was to deprecate support for implicit conversions from
ArrayRCP to ArrayView.  This was done to make this consistent with RCP which
does not provide an implicit conversion to Ptr.  The reason to disallow this
implicit conversion to to make the code more self documenting.  That way, if
you see a naked ArrayRCP passed to a function, you know that the function
takes the ArrayRCP and therefore is creating a persisting association.

Specific changes:
-----------------

(*) Teuchos_Array.hpp: Changed from arcp(...) to arcpFromArray(...) to be
consistent with arcpFromArrayView(...).

(*) Teuchos_ArrayRCPDecl.hpp: Deprecated implicit conversion from ArrayRCP to
ArrayView (see above).

(*) Sundance/.../TSFEpetraVectorSpace.cpp: Adding a missing header that was
hidden due to explicit instantiation.

Testing:
--------

I have enabled Sundance here even through it is now a SS code and does not
technically need to be tested before checkin.  However, given that Sundance is
a node dependency, this can't break PS code and Sundance has a strong
dependence on Thyra so I wanted to make sure that it still works.

=============================
Build/Test Cases Summary
=============================

Tue Dec 22 19:00:44 MST 2009

Enabled Packages: Sundance, RTOp, Teuchos, Thyra
Enabled all Forward Packages

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=1100,notpassed=0
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=1104,notpassed=0

Local commits for this build/test group:
----------------------------------------
65ba21f Added deprecated warnings support for RTOp and Thyra and refactored some code.
15a9d68 Changing the behavior of arrayView(p, size) to allow zero size input
6a60220 Added conversions from [const] Array<T> to ArrayRCP<[const] T> with unit tests