Convert windows testing to use git
Changes to allow kallikrates to pull from the git repository instead of the cvs
repository. This has been late in coming but there were some issues that had to
be worked out.

Apparently to properly find the git executable we have to give the full name of
the file and not just the name of the command. Since Windows uses extentions to
determine whether or not something is executable this conflicts with the
non-windows machines. Logic has been put in to find "git.cmd" on windows and
"git" on all other machines.

There is a known issue that ctest right now dies if you are starting with a
fresh checkout of the testing area. The work around right now is to run the
testing script once, then cd into the testing directory and clone the repository
manually. Updates to the repository seem to be working. It is unclear if this is
an issue with msysgit or with ctest. This same issue does not appear on anyother
machine though.