Fix complaint of an existing branch when doing release testing
eg by default creates local tracking branches for all remote branches when it
clones. This made checking out a release branch for testing easier. When we
switched to raw git that doesn't happen automatically. However the testing set
that up. Setting up a local branch that already exists causes and error though.

The solution is to try to switch to an existing local branch, if that fails then
create the local tracking branch and check it out. This is a little bit of a
shotgun approach, but it is cleaner in the code than inspecting the list of
local branches and remote branches, parsing that then creating (or not) any
local branches that are needed.