Fixed ambiguous overload for std::pow() error.
Calling pow(int,int) is ambiguous with gcc 4.1 compilers, which caused
numerous build errors in the nightly tests.  I added explicit casts
to double to fix this.  Note that the result is an int, so casting
to value_type doesn't make sense if value_type is not a scalar,
so I used double.