Allow non-owning RCPNode added first; enabling node tracing by default;
The previous RCPNodeTracer logic was to require that the first RCPNode be an
owning RCPNode and all other logic was based on this.  However, there were
cases in Tpetra where this was not true but were prefectly valid (but not as
safe perhaps).  The new logic has been generlized to allow non-owning RCPNode
objects to be added first.  Now the function
RCPNodeTracer::getExistingRCPNode(...) will only return an owning RCPNode
object which is the right behavior given how it is used to create a weak
[Array]RCP object to detect dangling references.  I added a unit test to test
the new behavior.  Note that if you add a non-owning RCPNode object first,
then it can not benefit from the dangling reference detection that it would of
had if the owning RCPNode was created first.

Here I am also turning on node tracing by default when Teuchos_ENABLE_DEBUG is
turned on.  This seems to only increase the cost of the MPI_DEBUG test suite
by a small amount for all the downstream packages from Teuchos.  Without node
tracing the total test time as 920 sec.  With node tracing enabled, the total
test time went up to 1106 sec.  This is a fairly small increase that brings
with it a great deal of extra checking. Interestingly, the MPI test suite for
Tpetra (which uses the Teucohs memory management classes at a very low level)
only went up from 15.75 sec to 16.96 sec.  Here are the timings of all the
tests on my machine gabriel:

0) MPI_DEBUG Results (without tracing):

  100% tests passed, 0 tests failed out of 1028

  Label Time Summary:
  Amesos         =  20.31 sec
  Anasazi        = 145.31 sec
  AztecOO        =   9.29 sec
  Belos          =  46.89 sec
  CTrilinos      =  40.47 sec
  EpetraExt      =   9.82 sec
  FEI            =  49.92 sec
  Galeri         =   1.47 sec
  GlobiPack      =   1.74 sec
  Ifpack         =  53.69 sec
  Intrepid       = 319.72 sec
  Isorropia      =  31.16 sec
  Kokkos         =   1.00 sec
  Komplex        =   1.08 sec
  ML             =  58.43 sec
  MOOCHO         = 156.21 sec
  Meros          =   0.86 sec
  NOX            = 203.53 sec
  OptiPack       =  12.91 sec
  RTOp           =  10.96 sec
  Rythmos        =  93.15 sec
  Sacado         = 235.95 sec
  Shards         =   2.87 sec
  Stratimikos    =  36.93 sec
  Teuchos        =  25.42 sec
  Thyra          =  36.84 sec
  Tpetra         =  15.75 sec

  Total Test time (real) = 920.81 sec

2) MPI_DEBUG_BOOST_TRACE Results (with node tracing):

  100% tests passed, 0 tests failed out of 1031

  Label Time Summary:
  Amesos         =  18.30 sec
  Anasazi        = 163.40 sec
  AztecOO        =   9.29 sec
  Belos          =  46.31 sec
  CTrilinos      =  57.20 sec
  EpetraExt      =   9.78 sec
  FEI            =  54.23 sec
  Galeri         =   0.33 sec
  GlobiPack      =   3.94 sec
  Ifpack         =  57.17 sec
  Intrepid       = 347.35 sec
  Isorropia      =  33.12 sec
  Kokkos         =   0.98 sec
  Komplex        =   1.08 sec
  ML             =  58.32 sec
  MOOCHO         = 214.64 sec
  Meros          =   0.61 sec
  NOX            = 293.00 sec
  OptiPack       =  14.66 sec
  RTOp           =  13.12 sec
  Rythmos        = 124.24 sec
  Sacado         = 241.47 sec
  Shards         =   0.62 sec
  Stratimikos    =  48.27 sec
  Teuchos        =  30.35 sec
  Thyra          =  40.83 sec
  Tpetra         =  16.96 sec

  Total Test time (real) = 1106.21 sec

Note that the differences in the number of tests 1028 and 1031 is due to boost
being enabled in the later which enables a few more tests.

=============================
Build/Test Cases Summary
=============================

Wed Dec 30 18:06:57 MST 2009

Enabled Packages: Teuchos, Thyra
Enabled all Forward Packages

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=1028,notpassed=0
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=1031,notpassed=0
2) MPI_DEBUG_BOOST_TRACE => passed: Trilinos/MPI_DEBUG_BOOST_TRACE: passed=1031,notpassed=0

Local commits for this build/test group:
----------------------------------------
1838a89 Allow non-owning RCPNode added first; enabling node tracing by default;
b28c5ac Adding some .gitignores for doxygen generated stuff.