Improved type of NOX::Abstract::Vector return arguments
NOX.Abstract.i now contains a typemap such that if a return argument
is NOX::Abstract::Vector, I attempt to downcast it to its derived
type.  Currently, the only possibility is NOX.Epetra.Vector, and this
in turn gets converted to an Epetra.Vector.  If the downcast fails,
then a NOX.Abstract.Vector is returned.