Added  Piro::Thyra::PerformAnalysis function.
This function perfoms analysis using MOOCHO, OptiPack
or Dakota (via TriKota). It takes a Thyra ModelEvaluator
(which can be a Piro Solver) and a parameter list.

A single test problem is checked in, which behaves
well for many hand-tested cases (parallel, adjusted
tolerances, etc.).

One known error remains: the OptiPack runs SegFault
during the destruction phase. Debugging statements
are in the code to identify the vector whose memory
is trying to be freed twice.