Changed alpha and beta in computePreconditioner method.
Since the same computePreconditioner call is made for
Jacobians and shiftedMatrix calls, this function was
incorrect to use  alpha=0, beta=1 all the time.
Now, alpha and beta are saved when one of these
two matrix computations are made, and used in the
computePreconditioner call. This statefullness is
admittedly fragile, but that is better then wrong.