Refactor to add a "PreconditionerFactory" into the hierarchy.  "BlockPreconditionerFactory" now inherits from "PreconditionerFactory"
Added new abstraction for unifying the creation and use of preconditioner factories.

The "BlockPreconditionerState" now inherits from the "PreconditionerState"

The PreconditionerState has most of the content, where as the
"BlockPreconditionerState" currently has no real functionality.
This is part of a refactor to move away from the purely "Block" abstractions
so that other nonblock preconditioner factories are easily implemented.

Drops the "BlockPreconditioner" abstraction in favor of the "Preconditioner" abstraction.

Now the Preconditioner factories all use the Preconditioner instread. This
will make it easier to make "BlockPreconditionerFactory" inherit from "PreconditionerFactory".

Rearranging the hierarchy so that PreconditionerFactory is the parent of BlockPreconditionerFactory

Also, propagated the change through out Teko, so that everything works and builds.
*** all tests passed ***

Fixed the static factory construction mechanism. It is know localized in "PreconditionerFactory"

This is the final commit in the refactoring. The idea was to enhance
the hierarchy so that nonblocked preconditioner factories could be
more easily created. This puts a "PreconditionerFactory" at the top
of the hierarchy, with the next class down being the "BlockPreconditionerFactory".

As of right now several outstanding issues remain:
   1. need to change Epetra functionality to take advantage of the new hierarchy
   2. the Add and Multiply preconditioners need to be "pushed up" the hierarchy to
      inherit from PreconditionerFactory
   3. examples needs to be modified so that it builds (possibly)
      + also build-preconditioner document needs to be updated to reflect this
   4. probably more testing of the new hierarchy would be beneficial