PyTrilinos updated to accommodate packages now supporting shared libs on Windows
It would appear that a methodical if incomplete attempt is being made
to get shared Trilinos libraries to work on Windows machines.  These
attempts appear to be complete for Teuchos, Epetra, Anasazi/Epetra and
NOX/Epetra at this time.  To get these modifications to work within
PyTrilinos, certain changes needed to be made when #include-ing or
%include-ing configuration files.  The changes in this commit refelect
the necessary changes to accommodate shared library support on
Windows.  Presumably, additional packages will undergo the same
procudure in the future, and PyTrilinos will need to be modified again
to support them.