Added downcast capability to NOX
Here I have merged changes made on the release 10 branch that add
downcasting of Epetra_Operators, NOX::Abstract::Group and
NOX::Epetra::LinearSystem so that python users get derived classes
rather than base classes.  The result is that the
NOX.Epetra.Interface.Jacobian.computeJacobian() method can be utilized
in python the way it was intended.  This is demonstrated in the
example script exNOX_2DSim.