Adding support for auto generation of client header files at configure time
It turns out that CMake does not consider macro defines when building the list
of include file dependencies for *.o files.  Therefore, we have to implement
the logic to configure the header files automatically with CMake.

The function also writes into the SIERRA/bjam/config_headers directory.  To
keep these scripts general I will have to remove this at some point and put in
a general argument EXTRA_WRITE_TO.

I changed the script create-explicit-instant-files.py to not create the
XXX.hpp by default because now it is configured.