Changes to support SG df/dp, dg/dp, dg/dx and dg/dx_dot OutArgs.
Added SG OutArgs for df/dp, dg/dp, dg/dx, and dg/dx_dot.  This required
a fair amount of changes to the SGModelEvaluator interface.  In
particular, SGModelEvaluator no longer replaces p and g for the
specified indices with block SG vectors, rather it adds new p and g
In/OutArgs for the block SG vectors.  Morever dg/dx and dg/dxdot have to
be used as operators instead of directly as multi-vectors for efficiency
(similar to how W is treated).  To suppor this, I added a simple wrapper
that gives the Epetra_Operator interface to an Epetra_MultiVector.
Finally I added the new OutArgs to the FEApp and setup the pce driver in
NOX to use them.