Removing explicit instant from Sun and adding it to GCC
I have removed explicit instantiation from the Sun build because I was seeing
link errors.  I am not sure if explicit instantation really works on this Sun
compiler.

However, I like explicit instantaition so I am changing over to explicit
instantiation for GCC.  Once this works, we can expand this to other
platforms.  Templates and explicit instantiation is the future of Trilinos so
we might as well blaze the path with Thyra ...