Additional fixes to windows build
Had to disable Tifpack on windows since it relies on Tpetra which is disabled
for the windows build.

The PATH was not set correctly for the shared build so the DLLs were not found
correctly for the shared build which on windows causes a segfault. Currently I
know of no better way to deal with this, but this is a serious issue with having
DLLs on windows.