Fix an exclude that was causing the framework tests to fail
rbgen is not part of the release so it was being excluded. However, this was
causing some of the framework tests to fail since the regex that excluded rbgen
was inadvertently also excluding rbgen from the static list of packages that
were set up solely for the framework testing. I made the rbgen exclude more
specific so that only the intended rbgen is removed from the tarball. I made
this same change to nearly every exclude statement since there was a slightly
possibility of this being an issue for other packages. There was minimal impact
to the list of files that were excluded from this change, which is good, but it
did catch one file that was inadvertently being excluded by one of the general
excludes.

In addition the aclocal.m4 file in Trilinos was also excluded, this file should
have been excluded before but was missed somehow.