Added logic to rewrite client template header if file has changed
The existing behavior of the script was to just see if the A.hpp file existed
and if it did, it would not write it again.  This did not work when switching
between implicit and explicit instantiation.  Now the CMake code will read the
existing file and if the contents are the same as what is going to be written,
it will not write the the file.  Otherwise, it will write the file.

I tested this manually for the following use cases:

1) Delete the file A.hpp and make sure it is created anew => passed

2) Reconfigure with an existing file A.hpp and make sure it is not rewirtten
=> passed (verified by looking at 'ls -l A.hpp').

3) Reconfigure with an existing file A.hpp which is not the same and make sure
it gets rewritten => passed (verified by looking at 'cat A.hpp' before and
after).

This is good to go.