updates to the find Trilinos capability
Add seperate variables for mpi libraries to the find Trilinos feature

While the mpi libraries were already being put in the list of tpl libraries mpi
is special in that you can't link different versions. This is being put in to
help client codes find the right mpi when they do a search for it. While this
could have been stripped out of the tpl library information it is much easier to
have it seperate, especially since Trilinos has it seperate to begin with.

Added a few new MPI variables that were requested for Tramonto so that it would
be easier to ensure the same testing environment could be used. These variables
were for making sure you can get the same mpi run command and use it properly.

Added version number support for find Trilinos capability

Adding the ability to look at and confirm compatibility with a given version
number when one is given to the cmake find_package command. This was only added
to the Trilinos variant of the find_package capability simply because each
package could have its own version number and aren't necessarily specified in
cmake files. Instead of being inconsistent with which packages support it, or
just using the Trilinos version number it seems best to consistently not support
it for packages.