This check in takes care of some build issues experienced with GCC 4.5.0.
This compiler has problems with an internal variable to the solver managers,
which is a type defined by a typedef.  When the internal variable is set,
using the type and not the typedef, the compiler believes the code is trying
to define two different variables.  I have not seen this issue with any
compiler previous to 4.5.0.  I have built several packages in Trilinos to
ensure that this issue isn't prevalent anywhere else.

I've built and tested this fix using GCC 4.5.0 and 4.2.2.