Adding checking for new unknown files early and failing with unit tests (Bug 4777)
It was discovered that the checkin-test.py script was not catching at the
beginning if there where new unknown files and it would not fail until the
final 'eg commit --amend' command was run (way too late).

To address this, I am changing the checkin-test.py script so that it runs 'eg
status' again and this time it greps the output looking for changed uncommitted
files and new unknown files.  I have good error messages now and good unit
tests in place for this.

I have hope I have captured all the ways a repo is not ready to be pushed
(unstaged changes, staged changes, new unknown files).  If there are other
ways the repo is not ready to be pushed, then I don't know what they are.

As for GeneralScriptSupport[_UnitTests].py, I added a helper function
isSubstrInMultiLineString(...) that is used to look for strings in the 'eg
status' output.

=============================
Build/Test Cases Summary
=============================

Thu May 13 10:20:02 MDT 2010

Enabled Packages: TrilinosFramework
Enabled all Forward Packages

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=52,notpassed=0 (2.04800775051 min)
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=52,notpassed=0 (1.1442367355 min)

Local commits for this build/test group:
----------------------------------------
c95c821 Adding checking for new unknown files early and failing with unit tests (Bug 4777)