Fix an escaping issue in cron_driver.py and add the necessary files for my mac
On macs the path that cmake is installed into has a space in it which the
cron_driver.py was not properly escaping before calling the ctest exe. I fixed
that by replacing all spaces with "\ " before we call the ctest exe.

added the files to make s909348 compatible with this new testing method. Note
that I have not enabled this testing method on s909348 yet.